% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getAccession}
\alias{getAccession}
\alias{getAccession,Hit-method}
\alias{getAccession,HitList-method}
\alias{getAccession,Iteration-method}
\alias{getAccession,IterationList-method}
\alias{getAccession,blastReportDB-method}
\title{Access components of Blast Hits}
\usage{
getAccession(x, ...)

\S4method{getAccession}{blastReportDB}(x, id, ...)

\S4method{getAccession}{Hit}(x)

\S4method{getAccession}{HitList}(x)

\S4method{getAccession}{Iteration}(x)

\S4method{getAccession}{IterationList}(x)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the accession numbers of hits.
}

