% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blastReport-class.r
\docType{class}
\name{BlastHeader-class}
\alias{BlastHeader-class}
\alias{new_BlastHeader}
\title{Class \code{"BlastHeader"}}
\description{
An S4 class that that serves as a container for BLAST header information:
}
\section{Slots}{

\describe{
\item{\code{version}}{<\code{character}>; version of BLAST used.}

\item{\code{reference}}{<\code{character}>; reference for BLAST.}

\item{\code{database}}{<\code{character}>; name of the database.}
}}
\examples{
showClass("BlastHeader")
}
\seealso{
\code{\linkS4class{BlastReport}}, \code{\linkS4class{BlastParameters}}
}
\keyword{internal}

