% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/blast.r
\name{blastp}
\alias{blastp}
\alias{blastp_short}
\title{Wrapper for the NCBI Protein-Protein BLAST}
\usage{
blastp(query, db = "nr", out = NULL, outfmt = "xml", max_hits = 20,
 evalue = 10, matrix = "BLOSUM62", remote = FALSE, ...)

blastp_short(query, db = "nr", out = NULL, outfmt = "xml", max_hits = 20,
 evalue = 10, matrix = "BLOSUM62", remote = FALSE, ...)
}
\arguments{
\item{...}{Additional parameters passed on to the BLAST commmand line
tools. See \href{http://www.ncbi.nlm.nih.gov/books/NBK1763/#CmdLineAppsManual.4_User_manual}{here}
for a description of common options.}

\item{query}{Query sequences as path to a FASTA file,
an \code{\linkS4class{XStringSet}} object, or a character vector.}

\item{db}{The database to BLAST against (default: nr).}

\item{out}{(optional) Output file for alignment.
If \code{NULL} and the BLAST result is returned as
a \code{\linkS4class{BlastReport}} or \code{\linkS4class{BlastTable}}
object.}

\item{outfmt}{Output format, \code{'xml'} or \code{'table'}.}

\item{max_hits}{How many hits to return (default: 20).}

\item{evalue}{Expect value cutoff (default: 10).}

\item{matrix}{Scoring matrix name (default: BLOSUM62).}

\item{remote}{Execute search remotely.}
}
\description{
\itemize{
\item{\code{blastp}} is the traditional BLASTP.
\item {\code{blastp_short}} is BLASTP optimised for residues shorter than 30.
}
}
\details{
Run \code{blastp()} without arguments to print usage and
arguments description.
}
\examples{
##
}
\seealso{
BLAST documentation at \url{http://www.ncbi.nlm.nih.gov/books/NBK1763/}.

Other blast functions: \code{\link{blastn}},
  \code{\link{blastn_short}}, \code{\link{blastn_short}},
  \code{\link{dc_megablast}}, \code{\link{dc_megablast}},
  \code{\link{megablast}}, \code{\link{megablast}};
  \code{\link{blastx}}; \code{\link{makeblasttdb}};
  \code{\link{qblast}}; \code{\link{rpsblast}};
  \code{\link{tblastn}}; \code{\link{tblastx}}
}

