% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blastReport-class.r
\docType{class}
\name{BlastParameters-class}
\alias{BlastParameters-class}
\alias{new_BlastParameters}
\title{Class \code{"BlastParameters"}}
\description{
An S4 class that that serves as the container for blast parameters and
statistics.
}
\section{Slots}{

\describe{
\item{\code{program}}{\code{character} vector; the BLAST flavour that generated the
data.}

\item{\code{matrix}}{\code{character} vector; name of the matrix (\code{NA} for
nucleotide blast).}

\item{\code{expect}}{\code{numeric} vector; cutoff value.}

\item{\code{penalties}}{\code{numeric} vector; open and extend penalties.}

\item{\code{sc_match}}{\code{integer} vector; match score for nucleotide-nucleotide
comparison.}

\item{\code{sc_mismatch}}{\code{integer} vector; mismatch penalty for nucleotide-
nucleotide comparison.}

\item{\code{filter}}{\code{character} vector; filter string.}

\item{\code{num_sequences}}{\code{character} vector; number of sequences in the
database.}

\item{\code{num_letters}}{\code{character} vector; number of letters in the database.}

\item{\code{hsp_length}}{\code{numeric} vector; effective HSP length.}

\item{\code{effective_space}}{\code{numeric} vector; effective search space.}

\item{\code{ka_params}}{\code{numeric} vector; kappa, lambda, entropy.}
}}
\examples{
showClass("BlastParameters")
}
\seealso{
\code{\linkS4class{BlastReport}}, \code{\linkS4class{BlastHeader}}
}
\keyword{internal}

