\docType{package}
\name{blastr}
\alias{blastr}
\alias{blastr-package}
\title{A wrapper arpound NCBI Blast.}
\description{
  Provide a set of classes and functions to manage NCBI
  Blast web service queries via R.
}
\details{
  \tabular{ll}{ Package: \tab blastr\cr Type: \tab
  Package\cr Version: \tab 0.0-2\cr Date: \tab
  2012-11-05\cr License: \tab MIT\cr }
}
\author{
  Gerhard Schoefl \email{gschof@yahoo.de}

  Maintainer: Gerhard Schoefl \email{gschof@yahoo.de}
}
\keyword{package}

