% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{class}
\name{BlastReport-class}
\alias{BlastReport-class}
\alias{new_BlastReport}
\alias{show,BlastReport-method}
\title{BlastReport}
\usage{
\S4method{show}{BlastReport}(object)
}
\description{
\code{"BlastReport"} is the top-level container for data parsed from NCBI
Blast XML output. It contains the component data classes:
\code{\linkS4class{BlastHeader}},
\code{\linkS4class{BlastParameters}}, and
\code{\linkS4class{IterationList}}.

The \code{\linkS4class{Iteration}} elements store results from individual
BLAST queries. Each \code{Iteration} holds a \code{\linkS4class{HitList}}
with possibly multiple \code{\linkS4class{Hit}s}, which, in turn, can
contain multiple high-scoring pairs (\code{\linkS4class{Hsp}s}).

Iterations, Hits, and Hsps can be extracted using the accessors
 \code{\link{getIteration}}, \code{\link{getHit}}, and \code{\link{getHsp}},
 or by directly subsetting a \code{blastReport} object.

E.g. \code{report[[1]][[1]]} will return the first hit in the first query.
}
\details{
The \code{show} methods for various blast objects can be modified by
a number of global options. Specifically, the number of
\code{\linkS4class{Hit}s} shown when displaying
\code{\linkS4class{Iteration}s} or \code{\linkS4class{HitList}s} are set
by \code{showHits}. Whether alignments are displayed or not is controlled
by \code{showAlignment}. Setting these options to \code{NULL}, restores
the defaults.

See the general documentation of \code{\link[methods]{show}} method for
the expected behavior.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: 
}}
\section{Slots}{

\describe{
\item{\code{header}}{Header information; \code{\linkS4class{BlastHeader}}.}

\item{\code{params}}{Blast parameters and statistics \code{\linkS4class{BlastParameters}}.}

\item{\code{iterations}}{Iterations; \code{\linkS4class{IterationList}}.}
}}
\examples{
# options("showHits" = 20)
# show(hitlist)
# options("showHits" = NULL)
# show(hitlist)
}
\seealso{
The constructor \code{\link{blastReport}}; the BLAST classes
 \code{\linkS4class{blastReportDB}} and \code{\linkS4class{blastTable}}

\code{\link[methods]{show}}
}

