\name{blastReportDB-class}
\alias{blastReportDB-class}
\alias{[,blastTable-method}
\alias{[[,blastTable-method}
\alias{$,blastTable-method}
\alias{names,blastTable-method}
\alias{show,blastTable-method}
\title{blastReportDB class}
\description{
  blastReportDB is an S4 class that provides a connection
  to a database of blast records organised in three tables:
  query, hit and hsp
}

