\name{blastTable}
\alias{blastTable}
\title{Parse NCBI BLAST table files into \linkS4class{blastTable} objects.}
\usage{
blastTable(blastfile)
}
\arguments{
  \item{blastfile}{Blast output in tabular format (file
  path or character vector)}
}
\value{
A \code{\linkS4class{blastTable}} object.
}
\description{
Parse NCBI BLAST table files into \linkS4class{blastTable}
objects.
}

