% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cmd.R
\name{check_cmd}
\alias{check_cmd}
\title{Check if Commands Are Available otherwise create conda environments for each tool}
\usage{
check_cmd(
  cmd = "blastn",
  env_name = "blast-env",
  verbose = "silent",
  force = FALSE
)
}
\description{
Check if Commands Are Available otherwise create conda environments for each tool
}
