% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gz.r
\name{compress_or_decompress}
\alias{compress_or_decompress}
\title{Compress or decompress files}
\usage{
compress_or_decompress(
  input,
  output,
  input_con,
  output_con,
  overwrite = FALSE,
  keep = FALSE
)
}
\arguments{
\item{input, output}{a file for input/output}

\item{input_con, output_con}{a function to create input/output connections}

\item{overwrite}{overwrite output if it exists}

\item{keep}{preserve input}
}
\description{
Compress or decompress input file and write the encoded byte-stream into output file.
}
\details{
The compression or decompression depends entirely on the passed \code{input_con} and \code{output_con}
functions. If \code{input_con} is \code{file()} and \code{output_con} a \code{gzfile()}, then the function performs
standard gzip compression.
}
\seealso{
Connections like \code{\link[base:connections]{base::file()}}, \code{\link[base:connections]{base::gzfile()}}, \code{\link[base:connections]{base::bzfile()}}, and \code{\link[base:connections]{base::xzfile()}}
that can be used by this function to work with gz, bz, or xz file compressions.
\code{\link[=gzip]{gzip()}} and \code{\link[=gunzip]{gunzip()}} which are simple user-friendly wrappers around this function.
}
