% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getQueryFrom}
\alias{getHitFrame}
\alias{getHitFrom}
\alias{getHitRange}
\alias{getHitTo}
\alias{getQueryFrame}
\alias{getQueryFrom}
\alias{getQueryRange}
\alias{getQueryTo}
\title{Access components of High-Scoring Pairs (Hsps)}
\usage{
getQueryFrom(x, max = FALSE)

getQueryTo(x, max = FALSE)

getQueryFrame(x, max = FALSE)

getHitFrom(x, max = FALSE)

getHitTo(x, max = FALSE)

getHitFrame(x, max = FALSE)

getQueryRange(x, max = FALSE)

getHitRange(x, max = FALSE)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}

\item{max}{If \code{TRUE} return data only for the hsp with the
highest bit score.}
}
\value{
An numeric vector or a list of numeric vectors.

An \linkS4class{IRanges} object.
}
\description{
Extract the position of the start, end, frame, or range of the residues of
query and hit sequences.
}

