\name{db_bulk_insert}
\alias{db_bulk_insert}
\title{Bulk insert a data.frame into a db table}
\usage{
  db_bulk_insert(con, tbl, df)
}
\arguments{
  \item{con}{a connection object.}

  \item{tbl}{name of table in database.}

  \item{df}{data.frame to insert}
}
\description{
  Bulk insert a data.frame into a db table
}
\keyword{internal}

