% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{db_info}
\alias{db_info}
\title{Metadata for an SQLite database.}
\usage{
db_info(con)
}
\arguments{
\item{con}{A connection object.}
}
\description{
Metadata for an SQLite database.
}
\keyword{internal}

