% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{db_list_fields}
\alias{db_list_fields}
\title{List available tables in an SQLite database}
\usage{
db_list_fields(x, tbl, ...)
}
\arguments{
\item{x}{A connection object.}

\item{tbl}{Name of the table.}

\item{...}{Additional arguments (Currently unused).}
}
\description{
List available tables in an SQLite database
}
\keyword{internal}

