% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{db_list_tables}
\alias{db_list_tables}
\title{List available tables in an SQLite database}
\usage{
db_list_tables(x, ...)
}
\arguments{
\item{x}{A connection object.}

\item{...}{Additional arguments (Currently unused).}
}
\description{
List available tables in an SQLite database
}
\keyword{internal}

