% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Defline-class}
\alias{Defline-class}
\title{Class \code{"Defline"}}
\description{
Class \code{"Defline"}
}
\section{Slots}{

\describe{
\item{\code{tag}}{}

\item{\code{accession}}{}

\item{\code{locus}}{}

\item{\code{description}}{}

\item{\code{species}}{}
}}
\examples{
showClass("Defline")
}
\keyword{classes}
\keyword{internal}

