% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DeflineSet-class}
\alias{DeflineSet-class}
\title{Class \code{"DeflineSet"} of \code{"Defline"} objects}
\description{
Class \code{"DeflineSet"} of \code{"Defline"} objects
}
\section{Slots}{

\describe{
\item{\code{.Data}}{Inherited from the \code{\link{list}} class.}
}}
\examples{
showClass("DeflineSet")
}
\keyword{classes}
\keyword{internal}

