\name{displayHTML}
\alias{displayHTML}
\title{display a html file in a browser}
\usage{
  displayHTML(html, browser = getOption("browser"),
    unlink = TRUE)
}
\arguments{
  \item{html}{file path or html encoded character string}

  \item{browser}{browser}

  \item{unlink}{remove temporary file}
}
\description{
  display a html file in a browser
}

