% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{do_log}
\alias{do_log}
\title{Log a message to stderr.}
\usage{
do_log(..., to = stderr())
}
\arguments{
\item{...}{Message strings.}

\item{to}{Where to write the log message to. If \code{NULL} no logging performed.}
}
\description{
Log a message to stderr.
}
\keyword{internal}

