% Generated by roxygen2 (4.0.2): do not edit by hand
\name{.blast}
\alias{.blast}
\title{Wrapper for the new NCBI BLAST+ tools}
\usage{
.blast(exec, query, db, outfmt = "xml", max_hits = 20, strand = "both",
  ..., intern = FALSE, show_cmd = FALSE, parse = TRUE)
}
\arguments{
\item{query}{Query sequences as path to a FASTA file,
an \code{\linkS4class{XStringSet}} object, or a character vector.}

\item{db}{Blast database name (defaults to 'nt' for blastn and 'nr' for
the rest).}

\item{outfmt}{XML or table (default: XML)}

\item{max_hits}{Maximum number of hits  to return. (defaults: 20)
Sets the '-max_target_seqs' parameter internally.}

\item{strand}{Query strand(s) to seach against database.}

\item{...}{Arguments passed on to the blast commmand line tools.}

\item{intern}{Set \code{TRUE} if no '-out' argument is specified.
Captures the blast output in an R character vector.}

\item{show_cmd}{If \code{TRUE} print the constructed command line
instead of passing it to \code{\link{system}}.}

\item{parse}{}

\item{program}{One of blastn, blastp, blastx, tblastn, or tblastx}
}
\description{
Wrapper for the new NCBI BLAST+ tools
}
\keyword{internal}

