% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blastr_cache.R
\name{get_blastr_cache}
\alias{get_blastr_cache}
\title{Retrieve System-Dependent Cache Path for BLASTr}
\usage{
get_blastr_cache(..., ext = NULL)
}
\arguments{
\item{...}{Additional path components to append to the cache directory.}

\item{ext}{An optional file extension (e.g., \code{"rds"}, \code{"csv"}) to
append to the final path.}
}
\value{
A path character string representing the path to the user cache
directory for the \code{BLASTr} package.
}
\description{
Determines the appropriate user cache directory for the \code{BLASTr} package
based on the operating system.
On macOS, it avoids using paths with spaces and follows the XDG base
directory specification.
}
\details{
This function uses \code{\link[tools:R_user_dir]{tools::R_user_dir()}} to determine the user cache
directory.
}
