% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getDefline}
\alias{getDefline}
\alias{getPrimaryDefline}
\title{Access components of Blast Hits}
\usage{
getDefline(x, ...)

getPrimaryDefline(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the (primary) definition lines.
}

