% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getGaps}
\alias{getGaps}
\title{Access components of High-Scoring Pairs (Hsps)}
\usage{
getGaps(x, max = FALSE)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}

\item{max}{If \code{TRUE} return data only for the hsp with the
highest bit score.}
}
\value{
An numeric vector or a list of numeric vectors
}
\description{
Extract the numbers of gaps.
}

