% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGeneID}
\alias{getGeneID}
\title{Access components of Blast Hits}
\usage{
getGeneID(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the GIs (gene identifiers) of hits.
}

