% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHit}
\alias{getHit}
\title{Access components of Blast Iterations (Queries)}
\usage{
getHit(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{IterationList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{hit} or \linkS4class{HitList} object.
}
\description{
Extract Blast hits.
}

