% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getHitCoverage}
\alias{getHitCoverage}
\title{Extract the total Hit Coverage.}
\usage{
getHitCoverage(x)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}
}
\description{
Caluculated as (alignment length - gaps) / hit length.
If hsps overlap, they are merged before caluclating the coverage.
}

