% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHitID}
\alias{getHitID}
\title{Access components of Blast Hits}
\usage{
getHitID(x, db = 'any', ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}

\item{db}{Database tag (e.g. 'gi', 'ref', 'gb', 'dbj', 'any')}
}
\value{
An character matrix or character vector.
}
\description{
Extract hit IDs.
}

