% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHitLen}
\alias{getHitLen}
\title{Access components of Blast Hits}
\usage{
getHitLen(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract the hit length.
}

