% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getHspHitID}
\alias{getHspHitID}
\title{Access components of Blast Hits}
\usage{
getHspHitID(x)
}
\arguments{
\item{x}{A \linkS4class{blastReportDB} connection.}
}
\value{
An numeric vector.
}
\description{
Extract Hit IDs for Hsps.
}

