% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getHspID}
\alias{getHspID}
\title{Access components of Blast Hits}
\usage{
getHspID(x)
}
\arguments{
\item{x}{A \linkS4class{BlastReportDB} connection.}
}
\value{
An numeric vector.
}
\description{
Extract Hsp IDs
}

