% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getHspNum}
\alias{getHspNum}
\title{Access components of High-Scoring Pairs (HSPs)}
\usage{
getHspNum(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}

\item{max}{If \code{TRUE} return index of the Hsp with the highest
bit score.}
}
\value{
An numeric vector or a list of numeric vectors.
}
\description{
Extract the HSP indices.
}

