% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getParams}
\alias{getParams}
\title{Extract parameters and statistics from a BLAST report}
\usage{
getParams(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{BlastReport}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{BlastParameters} object.
}
\description{
Extract parameters and statistics from a BLAST report
}

