% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getPercIdentity}
\alias{getMaxPercIdentity}
\alias{getPercIdentity}
\title{Access components of High-Scoring Pairs (HSPs)}
\usage{
getPercIdentity(x, max = FALSE)

getMaxPercIdentity(x)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}

\item{max}{If \code{TRUE} return data only for the hsp with the
highest bit score.}
}
\value{
An numeric vector or a list of numeric vectors
}
\description{
Extract Percent Identity, Percent Positives or Percent Gaps.
}

