% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQuery}
\alias{getQuery}
\title{Access components of Blast Iterations (Queries)}
\usage{
getQuery(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{BlastReportDB} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract a data frame with the fields:
query_id, query_def, query_len, n_hits
}

