% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQueryCoverage}
\alias{getQueryCoverage}
\title{Extract the total Query Coverage.}
\usage{
getQueryCoverage(x)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}
}
\description{
Caluculated as (alignment length - gaps) / query length.
If hsps overlap, they are merged before caluclating the coverage.
}

