% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQueryID}
\alias{getQueryID}
\title{Access components of Blast Iterations (Queries)}
\usage{
getQueryID(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{IterationList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract the query ID.
}

