% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQueryLen}
\alias{getQueryLen}
\title{Access components of BLAST Queries}
\usage{
getQueryLen(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{QueryList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract the query length.
}

