% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQueryNum}
\alias{getQueryNum}
\title{Access components of BLAST Queries}
\usage{
getQueryNum(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{QueryList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract the query (a.k.a. iteration) numbers.
}

