% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gz.r
\name{gzip}
\alias{gzip}
\alias{gunzip}
\alias{is_gzip}
\title{Compress and decompress using gzip}
\usage{
gzip(x, out = NULL, overwrite = FALSE, keep = FALSE)

gunzip(x, out = NULL, overwrite = FALSE, keep = FALSE)

is_gzip(x)
}
\arguments{
\item{x}{file to be compressed or decompressed using the \code{gz} algorithm}

\item{out}{\strong{optional} output file, see details}

\item{overwrite}{\strong{optional} overwrite existing file}

\item{keep}{\strong{optional} keep the original file}
}
\value{
the output path is silently returned
}
\description{
This is a native implementation of gzip using the \code{base::gzfile()} connection.
}
\details{
R can natively read and write in the \code{gz} format, but miss functions compress
and decompress files on disk. In this implementation, files are not being
read instead the compression happens on a chunked bytestream, so even
large files can be compressed or decompressed without exhausting RAM.

The interface tries to mimic the unix tools \code{gzip} and \code{gunzip}.
If \code{out} is not provided, the output file is derived from the input file, either
appending (for \code{gzip()}) or removing (for \code{gunzip()}) the \code{.gz} extension.
Once the file is successfully compressed/decompressed, the input file is removed
unless \code{keep = TRUE} is specified.
If for some reason the process fails, broken files are removed.

The \code{is_gzip()} is a simple utility function to check if the file is compressed
by gzip by comparing first two bytes of the file to a magic number "1f 8b".
}
\examples{
# Compression and decompression
test = tempfile()
writeLines("This is a test!", test)
test.gz = gzip(test)
file.exists(test) # no longer exists!
test = gunzip(test.gz)
file.exists(test.gz) # no longer exists!
readLines(test) # "This is a test!"

}
\seealso{
\code{\link[base:connections]{base::gzfile()}} for the gz connection
\code{\link[=compress_or_decompress]{compress_or_decompress()}} for the wokrhorse of \code{gzip} and \code{gunzip}
}
