% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{has_tables}
\alias{has_tables}
\title{Check if a SQLite database has specified tables}
\usage{
has_tables(con, tbl, ...)
}
\arguments{
\item{con}{a connection object}

\item{...}{}

\item{tables}{a character vector of table names}
}
\description{
Check if a SQLite database has specified tables
}
\keyword{internal}

