% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getHeader}
\alias{getHeader}
\title{Extract the header from a BLAST report}
\usage{
getHeader(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{blastReport}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{blastHeader} object.
}
\description{
Extract the header from a BLAST report
}

