% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getHitID}
\alias{getHitID}
\alias{getHitID,Hit-method}
\alias{getHitID,HitList-method}
\alias{getHitID,Iteration-method}
\alias{getHitID,blastReportDB-method}
\title{Access components of Blast Hits}
\usage{
getHitID(x, db = 'any', ...)

\S4method{getHitID}{blastReportDB}(x, id, ...)

\S4method{getHitID}{Hit}(x, db = "any")

\S4method{getHitID}{HitList}(x, db = "any")

\S4method{getHitID}{Iteration}(x, db = "any")
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}

\item{db}{Database tag (e.g. 'gi', 'ref', 'gb', 'dbj', 'any')}
}
\value{
An character matrix or character vector.
}
\description{
Extract hit IDs.
}

