\name{getHspHitID}
\alias{getHspHitID}
\alias{getHspHitID,blastReportDB-method}
\alias{getHspHitID-methods}
\title{Access components of Blast Hits}
\usage{
  getHspHitID(x)

  \S4method{getHspHitID}{blastReportDB}(x, ...)
}
\arguments{
  \item{x}{A \linkS4class{blastReportDB} connection.}
}
\value{
  An numeric vector.
}
\description{
  Extract Hit IDs for Hsps.
}

