\docType{methods}
\name{hspLen}
\alias{hspLen}
\alias{hspLen,Hit-method}
\alias{hspLen,HitList-method}
\alias{hspLen,Iteration-method}
\alias{hspLen,IterationList-method}
\title{Access components of Blast Hits}
\usage{
  hspLen(x)
}
\arguments{
  \item{x}{A \linkS4class{Hit} or \linkS4class{HitList}
  object.}
}
\value{
  An numeric vector.
}
\description{
  Extract the number of Hsps for each Hit.
}

