% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexed_fasta.r
\name{indexed_fasta}
\alias{indexed_fasta}
\alias{read_indexed_fasta}
\alias{[.indexed_fasta}
\alias{[[.indexed_fasta}
\alias{print.indexed_fasta}
\alias{length.indexed_fasta}
\alias{names.indexed_fasta}
\alias{names<-.indexed_fasta}
\title{Read indexed fasta}
\usage{
read_indexed_fasta(x, index = NULL)

\method{[}{indexed_fasta}(x, value)

\method{[[}{indexed_fasta}(x, value)

\method{print}{indexed_fasta}(x, ...)

\method{length}{indexed_fasta}(x)

\method{names}{indexed_fasta}(x)

\method{names}{indexed_fasta}(x) <- value
}
\arguments{
\item{x}{a fasta file}

\item{index}{an index to the corresponding fasta file}

\item{value}{in \code{[}, \code{[[}, or \code{$}, either a numeric index or a character string
which will be subsetted}

\item{...}{other arguments passed to or from functions}
}
\value{
an object of class \code{indexed_fasta}
}
\description{
Read a fasta file by using pre-computed index \code{.fai} using random access.
}
\details{
The index \code{.fai} is a precomputed file containing starts and lengths of sequences within
the fasta file. This allows obtaining specific sequences from the fasta file without
having to first read the whole file sequentially.

\code{read_indexed_fasta()} reads the index file and creates a \code{indexed_fasta} object.

The \code{indexed_fasta} object is internally represented as a list with a \code{file} the input
fasta and the index \code{index}. Functions such as \code{[}, \code{[[} and \code{names()} are defined
to make working with this object more pleasant and intutitive.
To view the internal structure of \code{obj} of class \code{indexed_fasta}, use \code{unclass(obj)}.

\code{[} is used to subset the \code{indexed_fasta} by subsetting the index.

\code{[[} is used to read one or more sequences from \code{indexed_fasta} using the \code{\link[base:seek]{base::seek()}}.
It returns object of the class \code{sequences} (see \code{blastr::sequences})

Methods \code{names()}, \verb{names<-()} allow getting or setting names,
\code{length()} allows for obtaining the number of sequences within the \code{indexed_fasta} object.
}
