% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_blast.R
\name{install_blast_conda}
\alias{install_blast_conda}
\title{Install BLAST+ CLI through conda}
\usage{
install_blast_conda(conda_env_name = "blastr-env")
}
\arguments{
\item{conda_env_name}{Name of the conda environment where
BLAST+ will be installed. Default is "blastr-env".}
}
\description{
Install BLAST+ command line tools in a conda environment.
Currently it don't works on ARM 64-bits MacOS systems.
}
