% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Iteration-class.r
\docType{class}
\name{Iteration-class}
\alias{Iteration-class}
\alias{getHit,Iteration-method}
\alias{new_Iteration}
\alias{nhits,Iteration-method}
\title{Class \code{"Iteration"}}
\usage{
\S4method{getHit}{Iteration}(x, i, drop = TRUE, ...)

\S4method{nhits}{Iteration}(x)
}
\description{
An S4 class that holds data parsed from an NCBI BLAST XML \sQuote{Iteration}
element. Each Iteration contains data about a query.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getHit}: Return a \code{\linkS4class{Hit}} or \code{\linkS4class{HitList}}

\item \code{nhits}: Returns the number of hits; <\code{numeric}>.
}}
\section{Slots}{

\describe{
\item{\code{iter_num}}{The number of the iteration; \code{"integer"}.}

\item{\code{query_id}}{Query identifier; \code{"character"}.}

\item{\code{query_def}}{Query definition; \code{"character"}.}

\item{\code{query_len}}{Query length; \code{"integer"}.}

\item{\code{hits}}{List of hits; \code{"\linkS4class{HitList}"}.}

\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}
}}
\examples{
showClass("Iteration")
}
\keyword{classes}

