% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getIteration}
\alias{getIteration}
\alias{getIteration,blastReport-method}
\title{Extract Iterations/Queries from a BLAST report}
\usage{
getIteration(x, ...)

\S4method{getIteration}{blastReport}(x, drop = TRUE)
}
\arguments{
\item{x}{A \code{\linkS4class{blastReport}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{Iteration} or \linkS4class{IterationList} object.
}
\description{
Extract Iterations/Queries from a BLAST report
}

