% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Iteration-class.r
\docType{class}
\name{IterationList-class}
\alias{IterationList-class}
\alias{getHit,IterationList-method}
\alias{nhits,IterationList-method}
\title{Class \code{"IterationList"} of \code{"Iteration"} objects}
\usage{
\S4method{getHit}{IterationList}(x, i, drop = TRUE, ...)

\S4method{nhits}{IterationList}(x)
}
\description{
An S4 class that holds a list of \code{\linkS4class{Iteration}} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getHit}: Returns a list of \code{\linkS4class{Hit}}s or \code{\linkS4class{HitList}}s

\item \code{nhits}: Returns the numbers of hits; <\code{numeric}>.
}}
\section{Slots}{

\describe{
\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}

\item{\code{.Data}}{Inherited from the \code{\link{list}} class.}
}}
\examples{
showClass("IterationList")
}
\seealso{
\code{\linkS4class{BlastReport}}
}
\keyword{classes}

