% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi.r
\name{ncbi}
\alias{ncbi}
\alias{ncbi_download}
\alias{ncbi_flatten}
\alias{ncbi_extract}
\title{Download, extract, and flatten the NCBI genome archives}
\usage{
ncbi(
  id,
  formats = NULL,
  dir = NULL,
  compress = TRUE,
  overwrite = FALSE,
  keep = FALSE
)

ncbi_download(
  id,
  dir = ".",
  formats = c("GENOME_FASTA", "GENOME_GTF"),
  overwrite = FALSE,
  quiet = TRUE,
  timeout = 3600
)

ncbi_flatten(
  archive,
  dir = NULL,
  formats = c("GENOME_FASTA", "GENOME_GTF"),
  keep = FALSE,
  overwrite = FALSE
)

ncbi_extract(archive, dir = NULL, keep = FALSE)
}
\arguments{
\item{id}{NCBI accession id of requested genome}

\item{formats}{requested file formats, such as \code{GENOME_FASTA} for fasta sequences
or \code{GENOME_GTF} for the annotation in the \code{.gtf} format.
For the \code{ncbi()} function, if \code{formats = NULL}, the available formats for the accession \code{id}
are obtained and all of them are downloaded. This might duplicate some information,
for instance the Genbank Flat File Format \code{gbff} (\code{GENOME_GBFF}) includes sequences
and annotation information, if you ar ealready downloading \code{GENOME_FASTA} and \code{GENOME_GTF},
you might consider to leave out this format to save space and bandwidth.
\code{ncbi_flatten()} need to have formats specified and defaults to \code{c("GENOME_FASTA", "GENOME_GTF")}.}

\item{dir}{\strong{optional} directory where the files will be downloaded. For \code{ncbi()},
\code{ncbi_flatten()}, and \code{ncbi_extract()}, the default folder is a folder with the same
name as the accession \code{id}. The \code{ncbi_download()} defaults to the current working directory.}

\item{compress}{\strong{optional} compress the extracted files using \code{\link[=gzip]{gzip()}}.}

\item{overwrite}{\strong{optional} overwrite existing files.}

\item{keep}{\strong{optional} keep the downloaded archive}

\item{quiet}{\strong{optional} print diagnostic information during download,
passed to \code{\link[=download.file]{download.file()}}}

\item{timeout}{\strong{optional} maximum allowed time for the connection in seconds,
passed to \code{\link[=download.file]{download.file()}}}

\item{archive}{a path to the downloaded NCBI genomic archive}
}
\description{
Functions to download, extract, and flatten the NCBI genome archives.
}
\details{
The NCBI genome archive allows you to download reference genomes through a web interface
or through REST API in a nice zip archive. There is only one issue, it comes in a deeply
nested structure with a lot of ancillary files.

The purpose of these functions is to download the NCBI archive, unpack it, extract
the files of interest, and rename them with a unified format with \verb{\{accession\}.\{extension\}}.

The \code{ncbi()} downloads, extract, and flattens the archive. If \code{formats = NULL},
it will perform an additional NCBI API request to determine which formats are available
for the requested genome.

\code{ncbi_download()} downloads a zip archive from the NCBI genome website.

\code{ncbi_extract()} is a simple wrapper around \code{\link[=unzip]{unzip()}} that extracts the NCBI genome archive.

\code{ncbi_flatten()} uses \code{ncbi_extract()} internally to flatten the archive. Flatteining
in here means extracting the requested files from a deeply nested structure and returning
a simple list of consistently named files.
}
