% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi.r
\docType{data}
\name{ncbi_formats}
\alias{ncbi_formats}
\alias{ncbi_available_formats}
\alias{ncbi_format_mapping_table}
\title{NCBI genomic formats}
\format{
An object of class \code{character} of length 8.
}
\usage{
ncbi_formats

ncbi_available_formats(id)

ncbi_format_mapping_table(id, assembly_name, formats = NULL)
}
\arguments{
\item{id}{NCBI genome accession id}

\item{assembly_name}{an assembly name obtained from parsing \code{assembly_data_report.jsonl}.
Normally, the assembly has the same name as the genome accession id, but in some cases it
might differ.}

\item{formats}{\strong{optional} requested formats to display}
}
\value{
In case of \code{ncbi_available_formats()}, a character vector with a subset of values
from \code{ncbi_formats}. The \code{ncbi_format_mapping_table()} returns a data.frame with original
and new file names.
}
\description{
Functions to work with NCBI genomic file formats.
}
\details{
The NCBI datasets API define file formats that are available to download together
with the genome. These formats carry additional information, such as coding sequences,
annotation, or translated protein sequences. However, not all formats are available for
all genomes. In the API documentation, they are refered to as annotation type, here
we refer to them as formats, but since most of these files are in the \code{fasta} format
and only \code{.gtf}, \code{.gff}, and \code{.gbff} files are considered annotation, neither term
is accurate.

The \code{ncbi_formats} object is a named vector of all possible formats.
The names are used in the NCBI JSON output, while the values are used in the API input,
as well as in the other ncbi functions.

The \code{ncbi_available_formats()} makes a NCBI API request to find what formats
are actually available for the requested genome.

The \code{ncbi_format_mapping_table()} is then used to translate the file names between
unflattened and flattened archive.
}
\keyword{datasets}
