\name{new_BlastHeader}
\alias{new_BlastHeader}
\title{BlastHeader-class}
\usage{
  new_BlastHeader(...)
}
\description{
  A container for blast header information:
}
\details{
  \describe{ \item{\code{version}:}{Version of BLAST used;
  \code{"character"}.} \item{\code{reference}:}{Reference
  for BLAST; \code{"character"}.}
  \item{\code{database}:}{Name of the database;
  \code{"character"}.} }
}
\keyword{internal}

