% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{nhits}
\alias{nhits}
\title{Access components of BLAST Queries}
\usage{
nhits(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{QueryList} object.}
}
\value{
An numeric vector.
}
\description{
Extract the number of Hits for each Query.
}

