% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nhsps}
\alias{nhsps}
\title{Access components of Blast Hits}
\usage{
nhsps(x)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}
}
\value{
An numeric vector.
}
\description{
Extract the number of Hsps for each Hit.
}

