% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthologs.r
\name{orthologs}
\alias{orthologs}
\title{Find orthologs based on RBH protein search.}
\usage{
orthologs(
  genes,
  query,
  subject,
  dir = ".",
  cache = ".cache/annotation.rds",
  nthreads = 1,
  keep = FALSE,
  type = c("RBH", "RBHS"),
  ...
)
}
\arguments{
\item{genes}{Gene names for which ortholog sequences should be found.}

\item{query, subject}{NCBI genome ID for query and subject respectively.}

\item{dir}{A directory where the NCBI genome data are downloaded,
the default is the current working directory.}

\item{cache}{A path to cache file. The default is ".cache/annotation.rds".
Set to \code{cache = NULL} to disable caching.}

\item{nthreads}{The number of threads parameter passed to NCBI BLAST+.
The default is \code{1}.}

\item{keep}{Keep the forward and backward (reciprocal) blast searches,}

\item{type}{Run the \code{RBH} or \code{RBHS} method, default is \code{RBH}
these are returned together with normal output in a named list}

\item{...}{Other arguments passed to \code{rblast} or \code{rsblast} respectively.}
}
\value{
A data.frame of with \code{length(genes)} rows and 5 columns:
\code{query_gene}, \code{query_protein}, \code{subject_gene}, \code{subject_protein}, and \code{subject_annotation}.
Where query gene couldn't be matched to query protein, or the query protein didn't rblast
any subject protein, the values are \code{NA_character_}.
}
\description{
This is a user-friendly function that downloads query and subject protein sequences,
and then performs a BLAST reciprocal best hit (RBH) search to identify ortholog sequences.
The input query and output subject genes are matched to proteins using the downloaded gtf
annotation files.
}
\details{
This is an example function that employs many parts of the \code{blastr} package.
First, the query and subject genomes are downloaded from the NCBI using the \code{\link[=ncbi]{ncbi()}} function.
The input gene names are matched to protein names using the query \code{gtf} file.
Then, RBH is run to identify matched subject protein sequences, which are then translated
to gene names using the subject \code{gtf} file.

Since parsing the \code{gtf} file might take some time, for repeated runs the parsed
\code{gtf} file can be cached.
}
\seealso{
\code{\link[=ncbi]{ncbi()}} for downloading the genomes from the NCBI genome database,
\code{\link[=read_fasta]{read_fasta()}} and \code{\link[=read_gtf]{read_gtf()}} for reading the \code{fasta} and \code{gtf} files
respectively,
\code{\link[=filecache]{filecache()}} for a simple on-disk caching,
\code{\link[=rblast]{rblast()}} and \code{\link[=rsblast]{rsblast()}} for the implementation of RBH and RBSH
}
