\docType{methods}
\name{getParams}
\alias{getParams}
\title{Extract BLAST parameters and statistics from a report}
\usage{
getParams(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{blastReport}} object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{blastParameters} object.
}
\description{
Extract BLAST parameters and statistics from a report
}

