% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta.R
\name{parse_fasta}
\alias{parse_fasta}
\title{Parse FASTA File into Sequences}
\usage{
parse_fasta(file_path)
}
\arguments{
\item{file_path}{Path to the FASTA file.}
}
\value{
A character vector containing the sequences.
}
\description{
Extract the sequences from a FASTA file, dropping
headers and joining multi-line records.
}
\examples{
fasta_path <- fs::path_package("BLASTr", "extdata", "minimal_db_blast", ext = "fasta")
seqs <- parse_fasta(file_path = fasta_path)

}
