\name{parseBlastTabular}
\alias{parseBlastTabular}
\title{read a blast hit table}
\usage{
  parseBlastTabular(blast_output)
}
\arguments{
  \item{blast_output}{Blast output in XML format (File or
  character vector)}
}
\value{
  A \code{\link{blastReport-class}} object.
}
\description{
  read a blast hit table
}

