\name{parseBlastXml}
\alias{parseBlastXml}
\title{Parse xml blast output}
\usage{
  parseBlastXml(blast_output)
}
\arguments{
  \item{blast_output}{Blast output in XML format (File or
  character vector)}
}
\value{
  A \code{\link{blastReport-class}} object.
}
\description{
  Parse xml blast output
}

