\name{parseDeflines}
\alias{parseDeflines}
\title{Parse fasta definition lines}
\usage{
  parseDeflines(defline, species = FALSE)
}
\arguments{
  \item{defline}{List or character vector of NCBI fasta
  deflines.}

  \item{species}{Parse out species designations.}
}
\description{
  Parse fasta definition lines
}

