% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query-class.r
\docType{class}
\name{Query-class}
\alias{Query-class}
\alias{getHit,Query-method}
\alias{getQueryDef,Query-method}
\alias{getQueryID,Query-method}
\alias{getQueryLen,Query-method}
\alias{getQueryNum,Query-method}
\alias{new_Query}
\alias{nhits,Query-method}
\alias{nhsps,Query-method}
\title{Class \code{"Query"}}
\usage{
\S4method{nhsps}{Query}(x)

\S4method{getHit}{Query}(x, i, drop = TRUE, ...)

\S4method{nhits}{Query}(x)

\S4method{getQueryNum}{Query}(x)

\S4method{getQueryID}{Query}(x)

\S4method{getQueryDef}{Query}(x)

\S4method{getQueryLen}{Query}(x)
}
\description{
An S4 class that holds data parsed from an NCBI BLAST XML \sQuote{iteration}
element. Each \code{"Query"} instance contains data from one query.
}
\section{Methods (by generic)}{
\itemize{
\item \code{nhsps}: Returns the number of HSPs for each hit; <\code{integer}>.

\item \code{getHit}: Return a \code{\linkS4class{Hit}} or \code{\linkS4class{HitList}}

\item \code{nhits}: Returns the number of hits; <\code{numeric}>.

\item \code{getQueryNum}: Returns the query (iteration) number; <\code{integer}>.

\item \code{getQueryID}: Returns the query ID; <\code{character}>.

\item \code{getQueryDef}: Returns the query definition; <\code{character}>.

\item \code{getQueryLen}: Returns the query length; <\code{integer}>.
}}
\section{Slots}{

\describe{
\item{\code{iter_num}}{The number of the queries (iterations); <\code{integer}>.}

\item{\code{query_id}}{Query identifier; \code{"character"}.}

\item{\code{query_def}}{Query definition; \code{"character"}.}

\item{\code{query_len}}{Query length; \code{"integer"}.}

\item{\code{hits}}{List of hits; \code{"\linkS4class{HitList}"}.}

\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}
}}
\examples{
showClass("Query")
}
\keyword{classes}

