\docType{methods}
\name{getQueryDef}
\alias{getQueryDef}
\alias{getQueryDef,Iteration-method}
\alias{getQueryDef,IterationList-method}
\alias{getQueryDef,blastReport-method}
\alias{getQueryDef,blastReportDb-method}
\title{Access components of Blast Iterations (Queries)}
\arguments{
  \item{x}{A \linkS4class{blastReport} or
  \linkS4class{IterationList} object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
  A character vector or list of character vectors.
}
\description{
  Extract the query definition.
}

