% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query-class.r
\docType{class}
\name{QueryList-class}
\alias{QueryList-class}
\alias{getHit,QueryList-method}
\alias{getQueryDef,QueryList-method}
\alias{getQueryID,QueryList-method}
\alias{getQueryLen,QueryList-method}
\alias{getQueryNum,QueryList-method}
\alias{nhits,QueryList-method}
\alias{nhsps,QueryList-method}
\title{Class \code{"QueryList"} of \code{"Query"} objects}
\usage{
\S4method{nhsps}{QueryList}(x)

\S4method{getHit}{QueryList}(x, i, drop = TRUE, ...)

\S4method{nhits}{QueryList}(x)

\S4method{getQueryNum}{QueryList}(x)

\S4method{getQueryID}{QueryList}(x)

\S4method{getQueryDef}{QueryList}(x)

\S4method{getQueryLen}{QueryList}(x)
}
\description{
An S4 class that holds a list of \code{\linkS4class{Query}} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{nhsps}: Returns the number of HSPs for each hit for each query;
<\code{list}> of <\code{integer}s>.

\item \code{getHit}: Returns a list of \code{\linkS4class{Hit}}s or \code{\linkS4class{HitList}}s

\item \code{nhits}: Returns the numbers of hits; <\code{list}> of <\code{numeric}s>.

\item \code{getQueryNum}: Returns the query (iteration) numbers; <\code{integer}>.

\item \code{getQueryID}: Returns the query IDs; <\code{character}>.

\item \code{getQueryDef}: Returns the query definitions; <\code{character}>.

\item \code{getQueryLen}: Returns the query lengths; <\code{integer}>.
}}
\section{Slots}{

\describe{
\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}

\item{\code{.Data}}{Inherited from the \code{\link{list}} class.}
}}
\examples{
showClass("QueryList")
}
\seealso{
\code{\linkS4class{BlastReport}}
}
\keyword{classes}

