% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast.r
\name{rblast}
\alias{rblast}
\alias{rsblast}
\title{Reciprocal BLAST}
\usage{
rblast(
  x,
  query,
  subject,
  type = c("blastn", "blastp"),
  query_db = NULL,
  subject_db = NULL,
  args = NULL,
  keep = FALSE,
  sort_by = "evalue",
  decreasing = FALSE,
  n = 5
)

rsblast(
  x,
  query,
  subject,
  type = c("blastn", "blastp"),
  query_db = NULL,
  subject_db = NULL,
  args = NULL,
  keep = FALSE,
  sort_by = "evalue",
  decreasing = FALSE,
  p = 0.1,
  n = 5
)
}
\arguments{
\item{x}{Names or ids of the sequences to be blasted, id is the first space-separated part
of a sequence name and is intepreted by the NCBI BLAST+ as a query name.}

\item{query, subject}{A character vector of sequences of the query and target organisms,
such as one from \code{\link[=read_fasta]{read_fasta()}}.}

\item{type}{a type of blast, either nucleotide \code{blastn} or protein \code{blastp} blast}

\item{query_db, subject_db}{pre-existing query and subject blast dabatases.
If not provided, they are created in a R's tempdir from the sequences, see \code{\link[=make_blast_db]{make_blast_db()}}.}

\item{args}{a character vector of additional arguments passed to blast.}

\item{keep}{Keep the forward and backward (reciprocal) blast searches,
these are returned together with normal output in a named list}

\item{sort_by}{A character vector of column names to sort by when chosing the best hit.
By default, the values are sorted by \code{evalue}. Other options to consider might be \code{bit_score}, or \code{perc_identity},
but \code{evalue} should already provide a comprehensive metric for the sequence match.}

\item{decreasing}{A logical vector ideally of the same length as \code{sort_by}.
Whether the columns specified by \code{sort_by} should be sorted in decreasing order.
Default for \code{evalue} is \code{FALSE} since smaller \code{evalue} is more significant.}

\item{n}{The number of subject returned for each blast search (\code{-max_target_seqs}), defaults to 5.}

\item{p}{The proportion of \code{bit_score} of the best hit within which other hits will be considered.
If \code{p = NULL}, standard RBH is run instead.}
}
\value{
if \code{keep = FALSE}, a data.frame summarizing the search with query name, maching
subject, percentage of their identity, alignment length, and number of mismatches.
Where subject wasn't identified, values are \code{NA}.
If \code{keep = TRUE}, a list containing the above data.frame as well as results from the forward
and backward searches, together with the best hits in both directions.
}
\description{
Perform BLAST reciprocal best hit (RBH) search.
}
\details{
BLAST reciprocal best hit search will first perform a forward search, take the best hits,
and then performs a backward search using these best hits. If the sequence used for
the forward search are identified as a best hit in the backward search, the matched
subject are reported.

The function \code{rblast} implements the standard reciprocal blast.
The function \code{rsblast} implements an extended version of reciprocal blast called
Reciprocal Best Similar Hit blast. In this method, the forward blast is performed as normal,
but in the backward (reciprocal) blast instead of taking only the best hit,
all hits that are similar enough (in terms of \code{bit_score}) to the best hit are considered.
This should outperform RBH in a situation where many orthologs exist.

Internally \code{rblast} is implemented within \code{rsblast} when \code{p = NULL}.
}
