% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQuerySeq}
\alias{getHitSeq}
\alias{getMatch}
\alias{getQuerySeq}
\title{Access components of High-Scoring Pairs (Hsps)}
\usage{
getQuerySeq(x, max = FALSE)

getHitSeq(x, max = FALSE)

getMatch(x, max = FALSE)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}

\item{max}{If \code{TRUE} return data only for the hsp with the
highest bit score.}
}
\value{
An \linkS4class{BString} or \linkS4class{BStringSet} object.
}
\description{
Extract the query or hit sequence.
}

