% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.r
\name{sequences}
\alias{sequences}
\alias{fasta}
\alias{read_fasta}
\alias{write_fasta}
\alias{is_sequences}
\alias{print.sequences}
\title{Read and write sequences in a fasta format}
\usage{
read_fasta(file)

write_fasta(x, file = "", nchar = 80)

is_sequences(x)

\method{print}{sequences}(x, ...)
}
\arguments{
\item{file}{input or output file with fasta sequences, can be gzipped.}

\item{x}{a named character vector assumed to be sequences.}

\item{nchar}{\strong{optional} a number of characters per line}

\item{...}{arguments passed to the print method (currently unused).}
}
\value{
\code{read_fasta} returns a named character vector of sequences with class
\verb{c("sequences", "character") allowing for type-checking. }write_fasta` doesn't have return value.
}
\description{
Read and write DNA, RNA, or protein sequences in a fasta format.
Files compressed with \code{gzip} (extension \code{gz}) are supported.
}
\details{
No checks are performed, malformed input might lead to malformed output.

\code{read_fasta()} and \code{write_fasta()} read and write sequences in a fasta format.

\code{is_sequences()} tests if the passed object are sequences. In R, operations
tend to typically not preserve attributes, like class. \code{in_sequences()} thus
rans several tests instead, but these are only heuristics. If you want to be sure,
reassign the class attribute such as with \code{attr(x, "class") = "sequences"}.
}
\examples{
seq1 = c("A" = "AAA", "B" = "BBB", "C" = "CCC")
fasta = tempfile(fileext = ".fasta")
writeLines(paste0(">", names(seq1), "\n", seq1), fasta)

seq2 = read_fasta(fasta)
identical(seq1, seq2)

}
