% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SQLiteDB-class}
\alias{SQLiteDB-class}
\alias{sqliteDB}
\title{Class \code{"SQLiteDB"}}
\description{
Class \code{"SQLiteDB"}
}
\section{Fields}{

\describe{
\item{\code{.con}}{An \code{\linkS4class{SQLiteConnection}}.}

\item{\code{.info}}{A list.}

\item{\code{.path}}{Path to the database.}
}}
\keyword{internal}

