% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blast.r
\name{update_blastdb}
\alias{update_blastdb}
\title{Wrapper for \code{update_blastdb}}
\usage{
update_blastdb(..., destdir = getOption("blastr.blastdb.path") \%||\% ".",
  decompress = TRUE, showall = FALSE, passive = FALSE, timeout = 120,
  force = FALSE)
}
\arguments{
\item{...}{Blast db to download}

\item{destdir}{Destination directory for databases.}

\item{decompress}{if \code{TRUE}, decompresses the archives in \code{destdir}
and deletes the downloaded archives.}

\item{showall}{if \code{TRUE}, show all available pre-formatted BLAST
databases.}

\item{passive}{Use passive FTP.}

\item{timeout}{Timeout on connection to NCBI (default: 120 seconds).}

\item{force}{Force download even if there is a archive already on local
directory.}
}
\description{
Download pre-formatted BLAST databases from NCBI ftp site.
}
\examples{
update_blastdb(showall = TRUE)
\dontrun{
update_blastdb("16SMicrobial", "human_genomic")
}
}

