% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.r
\name{blast}
\alias{blast}
\title{Run BLAST query}
\usage{
blast(query, subject = NULL, db = NULL, type = "blastn", out, outfmt)
}
\arguments{
\item{query}{a fasta file or a fasta object}

\item{subject}{\strong{optional} fasta file or a fasta object against which to blast, see details}

\item{db}{\strong{optional} a database created from \code{\link[=make_blast_db]{make_blast_db()}}, see details}

\item{type}{\strong{optional} type of BLAST to run, one of \code{blastn}, \code{blastp}, \code{blastx},
\code{tblastn} or \code{tblastx}.}

\item{out}{output path}

\item{outfmt}{output format}
}
\value{
dunno
}
\description{
Perform BLAST search
}
