% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{class}
\name{BlastHeader-class}
\alias{BlastHeader-class}
\alias{new_BlastHeader}
\title{BlastHeader}
\description{
A container for blast header information:
}
\section{Slots}{

\describe{
\item{\code{version}}{Version of BLAST used; \code{"character"}.}

\item{\code{reference}}{Reference for BLAST; \code{"character"}.}

\item{\code{database}}{Name of the database; \code{"character"}.}
}}
\seealso{
\code{\linkS4class{BlastReport}}, \code{\linkS4class{blastParameters}}
}
\keyword{internal}

