% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{BlastHeader-class}
\alias{BlastHeader-class}
\alias{new_BlastHeader}
\title{BlastHeader-class}
\description{
A container for blast header information:
}
\details{
\describe{
  \item{\code{version}:}{Version of BLAST used; \code{"character"}.}
  \item{\code{reference}:}{Reference for BLAST; \code{"character"}.}
  \item{\code{database}:}{Name of the database; \code{"character"}.}
}
}
\keyword{internal}

