% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLASTr-package.R
\docType{package}
\name{BLASTr-package}
\alias{BLASTr}
\alias{BLASTr-package}
\title{BLASTr: Run BLAST from R}
\description{
Run BLAST from R, with nice standards for ASV search. This package contains nice set of functions for identifying sequences from high throughput metagenomic experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heronoh/BLASTr}
  \item \url{https://heronoh.github.io/BLASTr/}
  \item Report bugs at \url{https://github.com/heronoh/BLASTr/issues}
}

}
\author{
\strong{Maintainer}: Heron Hilario \email{heronoh@gmail.com} (\href{https://orcid.org/0000-0002-2548-1053}{ORCID})

Authors:
\itemize{
  \item Queiroz Lucio Rezende \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID})
}

}
\keyword{internal}
