% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLASTr-package.R
\docType{package}
\name{BLASTr-package}
\alias{BLASTr}
\alias{BLASTr-package}
\title{BLASTr: Parallel taxonomic classification of metabarcoding sequences}
\description{
Seamlessly run BLAST searches within R, tailored for ASV (Amplicon Sequence Variant) analysis. This package offers a robust set of functions designed to efficiently identify and process sequences from high-throughput metagenomicmetagenomic experiments, following best practices for sequence data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heronoh/BLASTr}
  \item \url{https://heronoh.github.io/BLASTr/}
  \item Report bugs at \url{https://github.com/heronoh/BLASTr/issues}
}

}
\author{
\strong{Maintainer}: Heron Hilario \email{heronoh@gmail.com} (\href{https://orcid.org/0000-0002-2548-1053}{ORCID})

Authors:
\itemize{
  \item Lucio Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID})
}

}
\keyword{internal}
