% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blastReport-class.r
\docType{class}
\name{BlastReport-class}
\alias{BlastReport-class}
\alias{[,BlastReport-method}
\alias{[[,BlastReport-method}
\alias{getHeader,BlastReport-method}
\alias{getHit,BlastReport-method}
\alias{getParams,BlastReport-method}
\alias{getQuery,BlastReport-method}
\alias{getQueryDef,BlastReport-method}
\alias{getQueryID,BlastReport-method}
\alias{getQueryLen,BlastReport-method}
\alias{getQueryNum,BlastReport-method}
\alias{is.na,BlastReport-method}
\alias{new_BlastReport}
\alias{nhits,BlastReport-method}
\alias{show,BlastReport-method}
\title{Class \code{"BlastReport"}}
\usage{
\S4method{getHeader}{BlastReport}(x, ...)

\S4method{getParams}{BlastReport}(x)

\S4method{getQuery}{BlastReport}(x, i, drop = TRUE)

\S4method{getHit}{BlastReport}(x, i, drop = TRUE)

\S4method{nhits}{BlastReport}(x)

\S4method{getQueryNum}{BlastReport}(x)

\S4method{getQueryID}{BlastReport}(x)

\S4method{getQueryDef}{BlastReport}(x)

\S4method{getQueryLen}{BlastReport}(x)

\S4method{[}{BlastReport}(x, i, j, ..., drop = TRUE)

\S4method{[[}{BlastReport}(x, i, j, ...)

\S4method{is.na}{BlastReport}(x)

\S4method{show}{BlastReport}(object)
}
\description{
An S4 class that that serves as the top-level container for data parsed from
NCBI BLAST XML output. It contains the following top-level components:

\itemize{
   \item \code{\linkS4class{BlastHeader}}
   \item \code{\linkS4class{BlastParameters}}
   \item \code{\linkS4class{QueryList}}
}
}
\details{
The \code{\linkS4class{Query}} elements store results from individual
BLAST queries. Each \code{Query} holds a \code{\linkS4class{HitList}}
with possibly multiple \code{\linkS4class{Hit}s}, which, in turn, can
contain multiple high-scoring pairs (\code{\linkS4class{Hsp}s}).

Queries, Hits, and Hsps can be extracted using the accessors
\code{\link{getQuery}}, \code{\link{getHit}}, and \code{\link{getHsp}},
 or by directly subsetting a \code{BlastReport} object.

E.g. \code{report[[1]][[1]]} will return the first hit in the first query.

The \code{show} methods for various blast objects can be modified by
a number of global options. Specifically, the number of
\code{\linkS4class{Hit}s} shown when displaying
\code{\linkS4class{Query}s} or \code{\linkS4class{HitList}s} are set
by \code{showHits}. Whether alignments are displayed or not is controlled
by \code{showAlignment}. Setting these options to \code{NULL}, restores
the defaults.

See the general documentation of \code{\link[methods]{show}} method for
the expected behavior.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getHeader}: Return \code{\linkS4class{BlastHeader}}.

\item \code{getParams}: Return \code{\linkS4class{BlastParameters}}.

\item \code{getQuery}: Return \code{\linkS4class{Query}} or
\code{\linkS4class{QueryList}}.

\item \code{getHit}: Return a list of \code{\linkS4class{HitList}}s.

\item \code{nhits}: Returns the numbers of hits; <\code{numeric}>.

\item \code{getQueryNum}: Return query numbers; <\code{integer}>.

\item \code{getQueryID}: Return query IDs; <\code{character}>.

\item \code{getQueryDef}: Return query definitions; <\code{character}>.

\item \code{getQueryLen}: Return query lengths; <\code{integer}>.

\item \code{[}: Subset to return an \code{\linkS4class{QueryList}}.

\item \code{[[}: Subset to return an \code{\linkS4class{Query}}.

\item \code{is.na}: Indicate missing \code{\linkS4class{Query}}s.

\item \code{show}: 
}}
\section{Slots}{

\describe{
\item{\code{header}}{Header information; \code{\linkS4class{BlastHeader}}.}

\item{\code{params}}{Blast parameters and statistics \code{\linkS4class{BlastParameters}}.}

\item{\code{queries}}{Queries; \code{\linkS4class{QueryList}}.}
}}
\examples{
# options("showHits" = 20)
# show(hitlist)
# options("showHits" = NULL)
# show(hitlist)
}
\seealso{
The constructor \code{\link{blastReport}}; the BLAST classes
 \code{\linkS4class{BlastReportDB}} and \code{\linkS4class{BlastTable}}

\code{\link[methods]{show}}
}
\keyword{classes}

