% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blastReport}
\alias{blastReport}
\title{Parse NCBI BLAST XML files.}
\usage{
blastReport(x, asText = FALSE)
}
\arguments{
\item{x}{BLAST output in XML format (as file path or character string).}

\item{asText}{If \code{TRUE} the XML BLAST output is passed in as a string
instead of a file path.}
}
\value{
A \code{\linkS4class{BlastReport}} object.
}
\description{
Parse NCBI BLAST XML files.
}
\examples{
##
}

