% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{blastReport}
\alias{blastReport}
\title{Parse NCBI BLAST XML files.}
\usage{
blastReport(blast, asText = FALSE)
}
\arguments{
\item{blast}{Blast output in XML format (file path or character vector)}

\item{asText}{If \code{TRUE} the XML blast output is passed in as a string
instead of a file path.}
}
\value{
A \code{\linkS4class{BlastReport}} object.
}
\description{
Parse NCBI BLAST XML files.
}

