% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/blastReportDB-class.r
\docType{class}
\name{BlastReportDB-class}
\alias{BlastReportDB-class}
\alias{new_BlastReportDB}
\title{Class \code{"blastReportDB"}}
\description{
A reference calss that provides a connection to an SQLite database
containing BLAST records organised in three tables:
}
\details{
\bold{query} with fields:

\itemize{
   \item query_id     INTEGER    Primary key
   \item query_def    TEXT
   \item query_len    INTEGER
}

\bold{hit} with fields:

\itemize{
   \item query_id     INTEGER
   \item hit_id       INTEGER    Primary key
   \item hit_num      INTEGER
   \item gene_id      TEXT
   \item accession    TEXT
   \item definition   TEXT
   \item length       INTEGER
}

\bold{hsp} with fields:

\itemize{
   \item query_id     INTEGER
   \item hit_id       INTEGER
   \item hsp_id       INTEGER    Primary key
   \item hsp_num      INTEGER
   \item bit_score    FLOAT
   \item score        INTEGER
   \item evalue       FLOAT
   \item query_from   INTEGER
   \item query_to     INTEGER
   \item hit_from     INTEGER
   \item hit_to       INTEGER
   \item query_frame  INTEGER
   \item query_frame  INTEGER
   \item identity     INTEGER
   \item positive     INTEGER
   \item gaps         INTEGER
   \item align_len    INTEGER
   \item qseq         TEXT
   \item hseq         TEXT
   \item midline      TEXT
}
}
\seealso{
The constructors \code{\link{blastReportDB}} and \code{\link{blastReportDBConnect}},
 and the BLAST classes \code{\linkS4class{BlastReport}} and
 \code{\linkS4class{blastTable}}.
}

