% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blastTable}
\alias{blastTable}
\title{Parse NCBI BLAST table files into \linkS4class{BlastTable} objects.}
\usage{
blastTable(x)
}
\arguments{
\item{x}{BLAST output in tabular format (file path or character string)}
}
\value{
A \code{\linkS4class{blastTable}} object.
}
\description{
Parse NCBI BLAST table files into \linkS4class{BlastTable} objects.
}

