% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{db_bulk_insert}
\alias{db_bulk_insert}
\title{Insert a \code{data.frame} into a corresponding SQLite table}
\usage{
db_bulk_insert(x, tbl, df, ...)
}
\arguments{
\item{tbl}{Name of table in database.}

\item{df}{A \code{data.frame} matching \code{tbl}.}

\item{...}{Additional arguments (Currently \code{log}: Path to a log file).}

\item{con}{A connection object.}
}
\description{
Insert a \code{data.frame} into a corresponding SQLite table
}
\keyword{internal}

