% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{db_connect}
\alias{db_connect}
\title{Connect to an existing SQLite database.}
\usage{
db_connect(dbName = ":memory:", create = FALSE)
}
\arguments{
\item{dbName}{Path to an SQLite database [default: :memory:].}
}
\description{
Connect to an existing SQLite database.
}
\keyword{internal}
