\name{db_connect}
\alias{db_connect}
\title{Connect to an existing SQLite database.}
\usage{
  db_connect(dbName = ":memory:", create = FALSE)
}
\arguments{
  \item{dbName}{Path to an SQLite database [default:
  :memory:].}

  \item{create}{}
}
\description{
  Connect to an existing SQLite database.
}
\keyword{internal}

