\docType{methods}
\name{db_count}
\alias{db_count}
\title{Count rows in a db table}
\usage{
db_count(x, tbl)
}
\arguments{
  \item{x}{A connection object.}

  \item{tbl}{Name of table in database.}

  \item{...}{Additional arguments (Currently \code{log}:
  Path to a log file).}
}
\description{
Count rows in a db table
}
\keyword{internal}

