% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlite-db.r
\name{db_is_current}
\alias{db_is_current}
\title{Is a connection to an SQLite database still current.}
\usage{
db_is_current(con)
}
\arguments{
\item{con}{A connection object.}
}
\description{
Is a connection to an SQLite database still current.
}
\keyword{internal}

