\docType{methods}
\name{db_list_fields}
\alias{db_list_fields}
\title{List available tables in an SQLite database}
\arguments{
  \item{x}{A connection object.}

  \item{tbl}{Name of the table.}

  \item{...}{Additional arguments (Currently \code{log}:
  Path to a log file).}
}
\description{
  List available tables in an SQLite database
}
\keyword{internal}

