% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{db_list_tables}
\alias{db_list_tables}
\title{List available tables in an SQLite database}
\usage{
db_list_tables(x, ...)
}
\arguments{
\item{x}{A connection object.}

\item{...}{Additional arguments (Currently unused).}
}
\description{
List available tables in an SQLite database
}
\keyword{internal}

