\docType{methods}
\name{db_query}
\alias{db_query}
\title{Query an SQLite database.}
\arguments{
  \item{x}{A connection object.}

  \item{stmt}{An SQL statemant.}

  \item{j}{Subset the result.}

  \item{...}{Additional arguments (Currently \code{log}:
  Path to a log file).}
}
\description{
  Query an SQLite database.
}
\keyword{internal}

