\docType{methods}
\name{getDefline}
\alias{getDefline}
\alias{getDefline,Hit-method}
\alias{getDefline,HitList-method}
\alias{getDefline,Iteration-method}
\alias{getPrimaryDefline}
\alias{getPrimaryDefline,Hit-method}
\alias{getPrimaryDefline,HitList-method}
\alias{getPrimaryDefline,Iteration-method}
\alias{getPrimaryDefline,IterationList-method}
\title{Access components of Blast Hits}
\arguments{
  \item{x}{A \linkS4class{Hit} or \linkS4class{HitList}
  object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
  A character vector or list of character vectors.
}
\description{
  Extract the (primary) definition lines.
}

