% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit_codes.R
\name{exit_codes}
\alias{exit_codes}
\title{Retrieve Exit Codes and Standard Error from BLASTr Results}
\usage{
exit_codes(blast_res)
}
\arguments{
\item{blast_res}{A data frame containing the results of a BLAST search.}
}
\value{
A data frame with the exit codes and standard error messages.
}
\description{
This function extracts the exit codes and standard error messages from the
results of a BLAST search performed using the \code{\link[=parallel_blast]{parallel_blast()}} function.
}
