\docType{methods}
\name{getGeneID}
\alias{getGeneID}
\alias{getGeneID,Hit-method}
\alias{getGeneID,HitList-method}
\alias{getGeneID,Iteration-method}
\alias{getGeneID,IterationList-method}
\alias{getGeneID,blastReportDB-method}
\title{Access components of Blast Hits}
\arguments{
  \item{x}{A \linkS4class{Hit} or \linkS4class{HitList}
  object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
  A character vector or list of character vectors.
}
\description{
  Extract the GIs (gene identifiers) of hits.
}

