% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_results.R
\name{get_blast_results}
\alias{get_blast_results}
\title{Run BLASTn}
\usage{
get_blast_results(asv, num_thread, db_path, perc_id, perc_qcov_hsp)
}
\arguments{
\item{asv}{Sequence to be BLASTed}

\item{num_thread}{Number of threads to be used}

\item{db_path}{Path to BLAST database}

\item{perc_id}{Percent identity to be used}

\item{perc_qcov_hsp}{Percent query coverage to be used}
}
\value{
A \code{tibble} with the results of BLASTn
}
\description{
Retrieve BLAST results from a given ASV
}
