% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_results.R
\name{get_blast_results}
\alias{get_blast_results}
\title{Run BLASTn}
\usage{
get_blast_results(
  asv,
  db_path,
  num_threads,
  blast_type = "blastn",
  perc_id = 80,
  perc_qcov_hsp = 80,
  num_alignments = 4
)
}
\arguments{
\item{asv}{Vector of sequences to be BLASTed.}

\item{db_path}{Complete path do formatted BLAST database.}

\item{num_threads}{Number of threads to run BLAST on.
Passed on to BLAST+ argument \emph{-num_threads}.}

\item{blast_type}{One of the available BLAST+ search engines,
one of: "blastn", "blastp", "blastx", "tblastn", "tblastx".}

\item{perc_id}{Lowest identity percentage cutoff.
Passed on to BLAST+ \emph{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff.
Passed on to BLAST+ \emph{-qcov_hsp_perc}.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}
}
\value{
A \code{tibble} with the results of BLASTn for each sequence.
}
\description{
Retrieve BLAST results from a given ASV
}
\examples{
\dontrun{
blast_res <- BLASTr::get_blast_results(
  asv = "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAGAGTA
  CTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC",
  db_path = "/data/databases/nt/nt",
  perc_id = 80,
  num_thread = 1,
  perc_qcov_hsp = 80,
  num_alignments = 2
)
}
}
