% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_results.R
\name{get_blast_results}
\alias{get_blast_results}
\title{Run BLASTn}
\usage{
get_blast_results(
  asv,
  num_thread,
  total_cores,
  db_path,
  perc_ID,
  perc_qcov_hsp,
  num_alignments
)
}
\arguments{
\item{num_thread}{Number of threads to run BLAST on. Passed on to BLAST+ argument \emph{-num_threads}.}

\item{db_path}{Complete path do formatted BLAST databases.}

\item{perc_ID}{Lowest identity percentage cutoff. Passed on to BLAST+ \emph{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff. Passed on to BLAST+ \emph{-qcov_hsp_perc}.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}
}
\value{
A \code{tibble} with the results of BLASTn.
}
\description{
Retrieve BLAST results from a given ASV
}
\examples{
blast_res1 <- BLASTr::get_blast_results(asv = "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAGAGTACTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC",
                                        db_path = "/data/databases/nt/nt",
                                        perc_ID = 80,
                                        num_thread = 1,
                                        perc_qcov_hsp = 80,
                                        num_alignments = 2,
                                        blast_type = "blastn")

}
