% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_results.R
\name{get_blast_results}
\alias{get_blast_results}
\title{Run BLASTn}
\usage{
get_blast_results(
  asv,
  db_path,
  num_threads = 1L,
  blast_type = "blastn",
  perc_id = 80L,
  perc_qcov_hsp = 80L,
  num_alignments = 4L,
  verbose = c("silent", "cmd", "output", "full"),
  env_name = "blastr-blast-env"
)
}
\arguments{
\item{asv}{Vector of sequences to be BLASTed.}

\item{db_path}{Complete path do formatted BLAST database.}

\item{num_threads}{Number of threads to run BLAST on.
Passed on to BLAST+ argument \code{-num_threads}.}

\item{blast_type}{One of the available BLAST+ search engines,
#'   one of: \code{c("blastn", "blastp", "blastx", "tblastn", "tblastx")}.}

\item{perc_id}{Lowest identity percentage cutoff.
Passed on to BLAST+ \code{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff.
Passed on to BLAST+ \code{-qcov_hsp_perc}.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}

\item{verbose}{Should condathis::run() internal command be shown?}

\item{env_name}{The name of the conda environment with the parameter (i.e. "blastr-blast-env")}
}
\value{
A \code{tibble} with the results of BLASTn for each sequence.
}
\description{
Retrieve BLAST results from a given ASV
}
\examples{
\dontrun{
blast_res <- BLASTr::get_blast_results(
  asv = "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAGAGTA
  CTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC",
  db_path = "/data/databases/nt/nt",
  perc_id = 80,
  num_thread = 1,
  perc_qcov_hsp = 80,
  num_alignments = 2
)
}
}
