% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_header.R
\name{get_fasta_header}
\alias{get_fasta_header}
\title{Retrieve FASTA names from BLAST DB based on subjectIDs}
\usage{
get_fasta_header(id, db_path)
}
\arguments{
\item{id}{SubjectID from BLAST results}

\item{db_path}{Path to BLAST DB}
}
\description{
Retrieve FASTA names from BLAST DB based on subjectIDs
}
