% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_header.R
\name{get_fasta_header}
\alias{get_fasta_header}
\title{Get header from SubjectID}
\usage{
get_fasta_header(
  id,
  db_path,
  env_name = "blastr-blast-env",
  verbose = "silent"
)
}
\arguments{
\item{id}{SubjectID from BLAST results or any NCBI Nucleotide identifier.}

\item{db_path}{Complete path do formatted BLAST database.}

\item{env_name}{The name of the conda environment used to run
command-line tools (i.e. "blastr-blast-env").}

\item{verbose}{Should condathis::run() internal command be shown?}
}
\value{
Complete identifier for the SubjectID as is on the database.
}
\description{
Retrieve complete sequence header from BLAST DB based on
SubjectIDs
}
\examples{
\dontrun{
dna_fasta_path <- fs::path_package(
  "BLASTr", "extdata", "minimal_db_blast",
  ext = "fasta"
)
temp_db_path <- fs::path_temp("minimal_db_blast")
make_blast_db(fasta_path = dna_fasta_path, db_path = temp_db_path)
get_fasta_header(id = "AP011979.1", db_path = temp_db_path)
}
}
