% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tax_by_name.R
\name{get_tax_by_name}
\alias{get_tax_by_name}
\title{Get taxonomy ranks using \emph{NCBI Taxonomy Tax ID}}
\usage{
get_tax_by_name(
  organisms_names,
  parse_result = TRUE,
  verbose = FALSE,
  env_name = "entrez-env"
)
}
\arguments{
\item{organisms_names}{Vector of \emph{NCBI Taxonomy Tax ID} to retrieve taxonomy for.}

\item{parse_result}{Should the taxonomy be returned as the \code{efetch} returns it or should it be parsed into a tibble.}

\item{verbose}{Character string indicating whether to print verbose messages during the process. Default is \code{silent}.}

\item{env_name}{Character string specifying the name of the conda environment where \code{efetch} is installed. Default is \code{"blastr-entrez-env"}.}

\item{total_cores}{Number of threads to use. Defaults to 1.}

\item{show_command}{Should the \code{efetch} command on command line be printed for debugging. Defaults to TRUE.}
}
\value{
A tibble with all the taxonomic ranks for the corresponding taxID.
}
\description{
Recover complete taxonomy for a given \emph{NCBI Taxonomy Tax ID}
}
