% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tax_by_taxID.R
\name{get_tax_by_taxID}
\alias{get_tax_by_taxID}
\title{Retrieve Taxonomic Ranks Using NCBI Taxonomy Tax IDs}
\usage{
get_tax_by_taxID(
  organisms_taxIDs,
  parse_result = TRUE,
  verbose = FALSE,
  env_name = "entrez-env"
)
}
\arguments{
\item{organisms_taxIDs}{A character vector of NCBI Taxonomy Tax IDs for which to retrieve taxonomy information.}

\item{parse_result}{Logical indicating whether to parse the taxonomy information into a tibble (\code{TRUE}, default) or return the raw output as returned by \code{efetch} (\code{FALSE}).}

\item{verbose}{Logical indicating whether to print verbose messages during the process. Default is \code{FALSE}.}

\item{env_name}{Character string specifying the name of the conda environment where \code{efetch} is installed. Default is \code{"entrez-env"}.}
}
\value{
A tibble containing the taxonomic ranks for the given Tax IDs.
}
\description{
Retrieves complete taxonomy information for given NCBI Taxonomy Tax IDs by querying the NCBI database using the \code{efetch} command.
}
\examples{
\dontrun{
# Retrieve taxonomy for a single Tax ID
tax_info <- get_tax_by_taxID("9606") # Human

# Retrieve taxonomy for multiple Tax IDs
tax_ids <- c("9606", "10090", "10116") # Human, Mouse, Rat
tax_info <- get_tax_by_taxID(tax_ids)

# Get unparsed taxonomy result
raw_tax_info <- get_tax_by_taxID("9606", parse_result = FALSE)

# Enable verbose output
tax_info <- get_tax_by_taxID("9606", verbose = TRUE)
}
}
