% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getAccession}
\alias{getAccession}
\title{Access components of Blast Hits}
\usage{
getAccession(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the accession numbers of hits.
}

