% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDefline}
\alias{getDefline}
\alias{getPrimaryDefline}
\title{Access components of Blast Hits}
\usage{
getDefline(x, ...)

getPrimaryDefline(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the (primary) definition lines.
}

