% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getHeader}
\alias{getHeader}
\title{Extract the header of a BLAST report}
\usage{
getHeader(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{BlastReport}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{BlastHeader} object.
}
\description{
Extract the header of a BLAST report
}

