% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getHitCoverage}
\alias{getHitCoverage}
\title{Extract the total Hit Coverage.}
\usage{
getHitCoverage(x)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}
}
\description{
Caluculated as (alignment length - gaps) / hit length.
If hsps overlap, they are merged before caluclating the coverage.
}

