% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getHsp}
\alias{getHsp}
\title{Access components of Blast Hits}
\usage{
getHsp(x, ...)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
An \linkS4class{Hsp} or \linkS4class{HspList}.
}
\description{
Extract Hsps from a Hit.
}

