% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getHspHitID}
\alias{getHspHitID}
\title{Access components of BLAST Hits}
\usage{
getHspHitID(x)
}
\arguments{
\item{x}{A \linkS4class{BlastReportDB} connection.}
}
\value{
An numeric vector.
}
\description{
Extract Hit IDs for Hsps.
}

