% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getIteration}
\alias{getIteration}
\title{Extract Iterations/Queries from a BLAST report}
\usage{
getIteration(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{BlastReport}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{Iteration} or \linkS4class{IterationList} object.
}
\description{
Extract Iterations/Queries from a BLAST report
}

