% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIterNum}
\alias{getIterNum}
\title{Access components of Blast Iterations (Queries)}
\usage{
getIterNum(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{IterationList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract the Iteration (Query) numbers.
}

