% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQuery}
\alias{getQuery}
\title{Extract queries from a BLAST report}
\usage{
getQuery(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{BlastReport}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{Query} or \linkS4class{QueryList} object.
}
\description{
Extract queries from a BLAST report
}

