% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getQueryDef}
\alias{getQueryDef}
\title{Access components of Blast Iterations (Queries)}
\usage{
getQueryDef(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{IterationList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the query definition.
}

