% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getQueryDef}
\alias{getQueryDef}
\title{Access components of BLAST Queries}
\usage{
getQueryDef(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{QueryList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector or list of character vectors.
}
\description{
Extract the query definition.
}

