\docType{methods}
\name{getHeader}
\alias{getHeader}
\title{Extract the header from a BLAST report}
\usage{
getHeader(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{blastReport}} object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{blastHeader} object.
}
\description{
Extract the header from a BLAST report
}

