% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{class}
\name{Hit-class}
\alias{Hit-class}
\alias{new_Hit}
\title{An S4 class to represent a BLAST hit.}
\description{
A container for information parsed from an XML NCBI BLAST hit element.
Information about multiple Blast hits within a query is contained in a
\code{HitList}.
}
\section{Slots}{

\describe{
\item{\code{hit_num}}{The number of the hit; \code{"integer"}.}

\item{\code{hit_def}}{Hit definition; \code{"\linkS4class{DeflineSet}"}.}

\item{\code{hit_acc}}{Accession number; \code{"character"}.}

\item{\code{hit_len}}{Length of hit; \code{"integer"}.}

\item{\code{hsps}}{List of HSPs; \code{"\linkS4class{HspList}"}.}

\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}
}}
\seealso{
\code{"\linkS4class{BlastReport}"}
}

