\docType{methods}
\name{getHit}
\alias{getHit}
\alias{getHit,Iteration-method}
\alias{getHit,IterationList-method}
\alias{getHit,blastReport-method}
\title{Access components of Blast Iterations (Queries)}
\arguments{
  \item{x}{A \linkS4class{blastReport} or
  \linkS4class{IterationList} object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
  A \linkS4class{hit} or \linkS4class{HitList} object.
}
\description{
  Extract Blast hits.
}

