\docType{methods}
\name{getHitDef}
\alias{getHitDef}
\alias{getHitDef,Hit-method}
\alias{getHitDef,HitList-method}
\alias{getHitDef,Iteration-method}
\alias{getHitDef,blastReportDB-method}
\alias{getPrimaryHitDef}
\alias{getPrimaryHitDef,Hit-method}
\alias{getPrimaryHitDef,HitList-method}
\alias{getPrimaryHitDef,Iteration-method}
\alias{getPrimaryHitDef,IterationList-method}
\title{Access components of Blast Hits}
\arguments{
  \item{x}{A \linkS4class{Hit} or \linkS4class{HitList}
  object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
  A character vector or list of character vectors.
}
\description{
  Extract the (primary) hit definition.
}

