\docType{methods}
\name{hitLen}
\alias{hitLen}
\alias{hitLen,Iteration-method}
\alias{hspLen,IterationList-method}
\title{Access components of Blast Iterations (Queries)}
\usage{
  hitLen(x)
}
\arguments{
  \item{x}{A \linkS4class{blastReport} or
  \linkS4class{IterationList} object.}
}
\value{
  An numeric vector.
}
\description{
  Extract the number of Hits for each Iteration/Query.
}

