\docType{methods}
\name{getHitLen}
\alias{getHitLen}
\alias{getHitLen,Hit-method}
\alias{getHitLen,HitList-method}
\alias{getHitLen,Iteration-method}
\alias{getHitLen,IterationList-method}
\alias{getHitLen,blastReportDB-method}
\title{Access components of Blast Hits}
\arguments{
  \item{x}{A \linkS4class{Hit} or \linkS4class{HitList}
  object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
  A numeric vector or list of numeric vectors.
}
\description{
  Extract the hit length.
}

