% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Hit-class.r
\docType{class}
\name{HitList-class}
\alias{HitList-class}
\title{Class \code{"HitList"} of \code{"Hit"} objects}
\description{
An S4 class that holds a list of \code{\linkS4class{Hit}} objects.
}
\section{Slots}{

\describe{
\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}

\item{\code{.Data}}{Inherited from the \code{\link{list}} class.}
}}
\examples{
showClass("HitList")
}
\seealso{
\code{"\linkS4class{BlastReport}"}
}

