\docType{methods}
\name{getHspHitID}
\alias{getHspHitID}
\alias{getHspHitID,blastReportDB-method}
\title{Access components of Blast Hits}
\usage{
  getHspHitID(x)
}
\arguments{
  \item{x}{A \linkS4class{blastReportDB} connection.}
}
\value{
  An numeric vector.
}
\description{
  Extract Hit IDs for Hsps.
}

