% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getHspHitID}
\alias{getHspHitID}
\alias{getHspHitID,blastReportDB-method}
\title{Access components of Blast Hits}
\usage{
getHspHitID(x)

\S4method{getHspHitID}{blastReportDB}(x, id, max = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{blastReportDB} connection.}
}
\value{
An numeric vector.
}
\description{
Extract Hit IDs for Hsps.
}

