\docType{methods}
\name{getHspID}
\alias{getHspID}
\alias{getHspID,blastReportDB-method}
\title{Access components of Blast Hits}
\usage{
getHspID(x)

\S4method{getHspID}{blastReportDB}(x, id, max = FALSE, ...)
}
\arguments{
  \item{x}{A \linkS4class{blastReportDB} connection.}
}
\value{
An numeric vector.
}
\description{
Extract Hsp IDs
}

