% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{HspList-class}
\alias{HspList-class}
\title{Class \code{"HspList"} of \code{"Hsp"} objects}
\description{
An S4 class that holds a list of \code{\linkS4class{Hsp}} objects.
}
\section{Slots}{

\describe{
\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}

\item{\code{.Data}}{Inherited from the \code{\link{list}} class.}
}}
\examples{
showClass("HspList")
}
\seealso{
\code{"\linkS4class{BlastReport}"}
}

