\name{IterationList-class}
\alias{IterationList-class}
\title{IterationList-class}
\description{
  A list of \linkS4class{Iteration} objects.
}
\section{Slots}{
  \describe{ \item{\code{query_env}:}{Shared container for
  \code{query_id}, \code{query_def}, \code{query_len}, and
  \code{hit_len}; \code{"environment"}.}
  \item{\code{.Data}:}{Inherited from the
  \code{\link{list}} class.} }
}

