% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{class}
\name{IterationList-class}
\alias{IterationList-class}
\title{IterationList}
\description{
A list of \linkS4class{Iteration} objects.
}
\section{Slots}{

\describe{
\item{\code{query_env}}{Shared container for \code{query_id},
\code{query_def}, \code{query_len}, and \code{hit_len};
\code{"environment"}.}

\item{\code{.Data}}{Inherited from the \code{\link{list}} class.}
}}
\seealso{
\code{\linkS4class{blastReport}}
}

