% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blast_db.R
\name{make_blast_db}
\alias{make_blast_db}
\title{Make BLAST Database}
\usage{
make_blast_db(
  fasta_path,
  db_path,
  db_type = "nucl",
  taxid_map = NULL,
  parse_seqids = TRUE,
  verbose = "silent",
  env_name = "blastr-blast-env"
)
}
\arguments{
\item{fasta_path}{Path to the input FASTA file.}

\item{db_path}{Path to the output BLAST database.}

\item{db_type}{Type of database to create, either "nucl" or "prot".}

\item{taxid_map}{Optional path to a file mapping sequence IDs to taxonomy IDs.}

\item{parse_seqids}{Whether to parse sequence IDs.}

\item{verbose}{Should \verb{[condathis::run()]} internal command be shown?}

\item{env_name}{The name of the conda environment with the parameter (i.e. "blastr-blast-env")}
}
\value{
The result of the \code{makeblastdb} command.
}
\description{
Create a BLAST database from a FASTA file.
}
\examples{
\dontrun{
make_blast_db(
  fasta_path = fs::path_package(
    "BLASTr", "extdata", "minimal_db_blast",
    ext = "fasta"
  ),
  db_path = fs::path_temp("minimal_db_blast"),
  db_type = "nucl"
)
}

}
