% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blast_db.r
\name{make_blast_db}
\alias{make_blast_db}
\title{Make BLAST database}
\usage{
make_blast_db(x, out, type = c("nucl", "prot"), title = NULL, verbose = FALSE)
}
\arguments{
\item{x}{fasta file to make a database from, will be unzipped if required}

\item{out}{an output path and name of the database, such as \code{path/name.db}}

\item{type}{the type of database}

\item{title}{\strong{optional} BLAST database title, defaults to the input file name}

\item{verbose}{\strong{optional} prints diagnostic information}
}
\description{
Make BLAST database out of fasta file.
}
