% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blast.r
\name{makeblasttdb}
\alias{makeblasttdb}
\title{Wrapper for NCBI makeblastdb}
\usage{
makeblasttdb(input_file, input_type = "fasta", dbtype = "nucl", ...,
  show_log = TRUE, show_cmd = FALSE)
}
\arguments{
\item{input_file}{Input file/database name. Multiple file/database names
can be provided as a character vector.}

\item{input_type}{Type of data specified in input file. One of \dQuote{fasta},
\dQuote{blastdb}, \dQuote{asn1_bin}, or \dQuote{asn1_txt}.}

\item{dbtype}{Molecule type of target db (\dQuote{nucl} or \dQuote{prot}).}

\item{...}{further arguments passed to \code{makeblastdb}.}

\item{show_log}{print log file.}

\item{show_cmd}{print the command line instead of executing it.}
}
\description{
Generate a BLAST database from FASTA files.
}
\examples{
## Retrieve, for instance, the human reference genome files located at
## ftp.ncbi.nlm.nih.gov/genomes/H_sapiens/Assembled_chromosomes/seq/
## and generate an input file hs_ref.fa by inflating and concatenating
## the hs_ref_*.fa.gz files.
\dontrun{
makeblasttdb("hs_ref.fa", input_type = "fasta", dbtype = "nucl",
  parse_seqids = TRUE, out = "hs_ref", title = "Human reference assembly, GRCh28")
}
}
\seealso{
Other blast functions: \code{\link{blastn}},
  \code{\link{blastn_short}}, \code{\link{blastn_short}},
  \code{\link{dc_megablast}}, \code{\link{dc_megablast}},
  \code{\link{megablast}}, \code{\link{megablast}};
  \code{\link{blastp}}, \code{\link{blastp_short}},
  \code{\link{blastp_short}}; \code{\link{blastx}};
  \code{\link{qblast}}; \code{\link{rpsblast}};
  \code{\link{tblastn}}; \code{\link{tblastx}}
}

