\docType{methods}
\name{nhits}
\alias{nhits}
\alias{nhits,Iteration-method}
\alias{nhits,IterationList-method}
\title{Access components of Blast Iterations (Queries)}
\usage{
  nhits(x)
}
\arguments{
  \item{x}{A \linkS4class{blastReport} or
  \linkS4class{IterationList} object.}
}
\value{
  An numeric vector.
}
\description{
  Extract the number of Hits for each Iteration/Query.
}

