% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{nhits}
\alias{nhits}
\title{Access components of Blast Iterations (Queries)}
\usage{
nhits(x)
}
\arguments{
\item{x}{A \linkS4class{blastReport} or \linkS4class{IterationList} object.}
}
\value{
An numeric vector.
}
\description{
Extract the number of Hits for each Iteration/Query.
}

