% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{nhits}
\alias{nhits}
\title{Access components of Blast Iterations (Queries)}
\usage{
nhits(x, ...)
}
\arguments{
\item{x}{A \linkS4class{BlastReport} or \linkS4class{IterationList} object.}
}
\value{
An numeric vector.
}
\description{
Extract the number of Hits for each Iteration/Query.
}

