\docType{methods}
\name{nhsps}
\alias{nhsps}
\alias{nhsps,Hit-method}
\alias{nhsps,HitList-method}
\alias{nhsps,Iteration-method}
\alias{nhsps,IterationList-method}
\title{Access components of Blast Hits}
\usage{
  nhsps(x)
}
\arguments{
  \item{x}{A \linkS4class{Hit} or \linkS4class{HitList}
  object.}
}
\value{
  An numeric vector.
}
\description{
  Extract the number of Hsps for each Hit.
}

