% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.r
\name{nhsps}
\alias{nhsps}
\title{Access components of BLAST Hits}
\usage{
nhsps(x)
}
\arguments{
\item{x}{A \linkS4class{Hit} or \linkS4class{HitList} object.}
}
\value{
An numeric vector.
}
\description{
Extract the number of HSPs for each Hit.
}

