% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_blast.R
\name{parallel_blast}
\alias{parallel_blast}
\title{Run Parallel BLAST}
\usage{
parallel_blast(
  asvs,
  db_path,
  out_file,
  out_rds,
  num_thread,
  perc_id,
  perc_qcov_hsp
)
}
\arguments{
\item{asvs}{Vector of sequences to be BLASTed}

\item{db_path}{Path to BLAST database}

\item{out_file}{Path to output file}

\item{out_rds}{Path to output RDS file}

\item{num_thread}{Number of threads to be used}

\item{perc_id}{Percent identity to be used}

\item{perc_qcov_hsp}{Percent query coverage to be used}
}
\description{
Run parallel BLAST for set of sequences
}
