% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_blast.R
\name{parallel_blast}
\alias{parallel_blast}
\title{Run Parallelized BLAST}
\usage{
parallel_blast(
  query_seqs,
  db_path,
  ...,
  total_cores = 1L,
  num_threads = 1L,
  blast_type = "blastn",
  perc_id = 80L,
  perc_qcov_hsp = 80L,
  num_alignments = 4L,
  retry_times = 10L,
  mt_mode = c("2", "1", "0"),
  verbose = c("progress", "silent", "cmd", "output", "full"),
  env_name = "blastr-blast-env",
  asvs = deprecated(),
  out_file = deprecated(),
  out_RDS = deprecated()
)
}
\arguments{
\item{query_seqs}{Character vector with sequences.}

\item{db_path}{Path to the formatted BLAST database.}

\item{...}{These dots are for future extensions and must be empty.}

\item{total_cores}{Number of parallel BLAST processes to run.}

\item{num_threads}{Number of threads/cores to run each BLAST process on.}

\item{blast_type}{BLAST+ executable to be used on search.}

\item{perc_id}{Lowest identity percentage cutoff.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff.}

\item{num_alignments}{Number of alignments to retrieve results for each query
sequence. Defaults to \code{4}.}

\item{retry_times}{Number of times to retry failed BLAST jobs.
Defaults to \code{10}.}

\item{mt_mode}{Multithreading mode to be used by BLAST+.
One of: \code{c("2", "1", "0")}. See BLAST+ manual for details.}

\item{verbose}{Strategy used for showing outputting internal commands.
Defaults to "progress".}

\item{env_name}{The name of the conda environment used to run
command-line tools. Defaults to \code{"blastr-blast-env"}.}

\item{asvs}{Deprecated. Same as \code{query_seqs}. Use \code{query_seqs} instead.}

\item{out_file}{Deprecated. Path to output \code{.csv} file on an existing directory.}

\item{out_RDS}{Deprecated. Path to output \code{RDS} file on an existing directory.}
}
\value{
A tibble with the BLAST tabular output.
}
\description{
Run parallel BLAST for a set of sequences
}
\examples{
\dontrun{
dna_fasta_path <- fs::path_package(
  "BLASTr", "extdata", "minimal_db_blast",
  ext = "fasta"
)
temp_db_path <- fs::path_temp("minimal_db_blast")
make_blast_db(fasta_path = dna_fasta_path, db_path = temp_db_path)
asvs_string <- c(
  "CTAGCCATAAACTTAAATGAAGCTATACTAA",
  "ACTCGTTCGCCAGAGTACTACAAGCGAAAG"
)
blast_res <- parallel_blast(
  query_seqs = asvs_string,
  db_path = temp_db_path
)
blast_res
}

}
