% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_blast.R
\name{parallel_blast}
\alias{parallel_blast}
\title{Run BLAST on parallel}
\usage{
parallel_blast(
  asvs,
  db_path,
  out_file,
  out_RDS,
  num_threads,
  total_cores,
  perc_ID,
  perc_qcov_hsp,
  num_alignments
)
}
\arguments{
\item{asvs}{Vector of sequences to be BLASTed.}

\item{db_path}{Complete path do formatted BLAST database.}

\item{out_file}{Complete path to output file}

\item{out_RDS}{Complete path to output RDS file.}

\item{num_threads}{Number of threads to run BLAST on. Passed on to BLAST+ argument \emph{-num_threads}.}

\item{total_cores}{Total available cores to parallelize BLAST. Chech your max with \emph{future::availableCores()}}

\item{perc_ID}{Lowest identity percentage cutoff. Passed on to BLAST+ \emph{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff. Passed on to BLAST+ \emph{-qcov_hsp_perc}.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}

\item{blast_type}{BLAST+ executable to be used on search.}
}
\value{
A tibble with the BLAST tabular output.
}
\description{
Run parallel BLAST for set of sequences
}
\examples{
blast_res <- BLASTr::parallel_blast(asvs = ASVs_test[1],
                                 db_path = "/data/databases/nt/nt",
                                 out_file = "~/blast_out.csv",
                                 out_RDS = "~/blast_out.RDS",
                                 total_cores = 77,
                                 perc_ID = 80,
                                 num_thread = 1,
                                 perc_qcov_hsp = 80,
                                 num_alignments = 4)

}
