% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_get_tax.R
\name{parallel_get_tax}
\alias{parallel_get_tax}
\title{Get taxonomy ranks for a list of \emph{NCBI Taxonomy Tax ID}}
\usage{
parallel_get_tax(organisms_taxIDs, parse_result = TRUE, total_cores = 1)
}
\arguments{
\item{parse_result}{Should the taxonomy be returned as the \emph{efetch} returns it or should it be parsed into a tibble.}

\item{total_cores}{Number of threads to use. Defaults to 1.}

\item{taxIDs}{Vector of \emph{NCBI Taxonomy Tax ID} to retrieve taxonomy for.}

\item{perc_id}{Lowest identity percentage cutoff.
Passed on to BLAST+ \emph{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff.
Passed on to BLAST+ \emph{-qcov_hsp_perc}.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}

\item{blast_type}{One of the available BLAST+ search engines,
one of: "blastn", "blastp", "blastx", "tblastn", "tblastx".}
}
\value{
Unformatted BLAST results.
For results formatted as tibble, please use \code{BLASTr::get_blast_results()}
}
\description{
Get taxonomy ranks for a list of \emph{NCBI Taxonomy Tax ID} in parallel.
}
\examples{
blast_res <- BLASTr::run_blast(
  asv = "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAG
    AGTACTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC",
  db_path = "/data/databases/nt/nt",
  perc_id = 80,
  num_thread = 1,
  perc_qcov_hsp = 80,
  num_alignments = 2
)

}
