% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_get_tax.R
\name{parallel_get_tax}
\alias{parallel_get_tax}
\title{Get taxonomy ranks for a list of \emph{NCBI Taxonomy Tax ID}}
\usage{
parallel_get_tax(organisms_taxIDs, parse_result = TRUE, total_cores = 1)
}
\arguments{
\item{organisms_taxIDs}{Vector of \emph{NCBI Taxonomy Tax ID} to retrieve taxonomy for.}

\item{parse_result}{Should the taxonomy be returned as the \emph{efetch} returns it or should it be parsed into a tibble.}

\item{total_cores}{Number of threads to use. Defaults to 1.}
}
\value{
Vector of Tax Ids.
}
\description{
Get taxonomy ranks for a list of \emph{NCBI Taxonomy Tax ID} in parallel.
}
