% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta.R
\name{parse_fasta}
\alias{parse_fasta}
\title{Extract Sequences From FASTA file}
\usage{
parse_fasta(file_path)
}
\arguments{
\item{file_path}{Path to FASTA file.}
}
\value{
vector containing the sequences.
}
\description{
Extract nucleotide or aminoacid sequences from FASTA file.
}
