% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getQueryLen}
\alias{getQueryLen}
\alias{getQueryLen,Iteration-method}
\alias{getQueryLen,IterationList-method}
\alias{getQueryLen,blastReport-method}
\alias{getQueryLen,blastReportDB-method}
\title{Access components of Blast Iterations (Queries)}
\usage{
getQueryLen(x, ...)

\S4method{getQueryLen}{blastReport}(x)

\S4method{getQueryLen}{blastReportDB}(x, id, ...)

\S4method{getQueryLen}{Iteration}(x)

\S4method{getQueryLen}{IterationList}(x)
}
\arguments{
\item{x}{A \linkS4class{blastReport} or \linkS4class{IterationList} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector or list of numeric vectors.
}
\description{
Extract the query length.
}

