% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast.r
\name{rblast}
\alias{rblast}
\title{Reciprocal BLAST}
\usage{
rblast(
  x,
  query,
  subject,
  type = c("blastn", "blastp"),
  dir = ".",
  keep = FALSE
)
}
\arguments{
\item{x}{id or names of the sequences to be blasted, id is the first space-separated part
of the sequence name and is intepreted by blast as query name}

\item{query, subject}{sequences of the query and target organisms}

\item{type}{a type of blast, either nucleotide \code{blastn} or protein \code{blastp} blast}

\item{dir}{= "." \strong{optional} wokring directory}

\item{keep}{\strong{optional} keep blast searches, these are returned together with normal output
in a list}
}
\value{
if \code{keep = FALSE}, a data.frame summarizing the search with query name, maching
subject, percentage of their identity, alignment length, and number of mismatches.
Where subject wasn't identified, values are \code{NA}.
If \code{keep = TRUE}, a list containing the above data.frame as well as results from the forward
and backward searches, together with the best hits in both directions.
}
\description{
Perform BLAST reciprocal best hit (RBH) search.
}
\details{
BLAST reciprocal best hit search will first perform a forward search, take the best hits,
and then performs a backward search using these best hits. If the sequence used for
the forward search are identified as a best hit in the backward search, the matched
subject are reported.
}
