% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run BLAST}
\usage{
run_blast(asv, db_path, num_alignments = 3, num_thread, perc_id, perc_qcov_hsp)
}
\arguments{
\item{asv}{Sequence to be BLASTed}

\item{db_path}{Path to BLAST database}

\item{num_alignments}{Number of alignments to be returned.
Default: the 3 best alignments.}

\item{num_thread}{Number of threads to be used}

\item{perc_id}{Percent identity to be used}

\item{perc_qcov_hsp}{Percent query coverage to be used}
}
\description{
Run BLAST for a single sequence
}
