% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run BLAST}
\usage{
run_blast(
  asv,
  db_path,
  num_alignments = 4L,
  num_threads = 1L,
  blast_type = "blastn",
  perc_id = 80L,
  perc_qcov_hsp = 80L,
  verbose = c("silent", "cmd", "output", "full"),
  env_name = "blastr-blast-env"
)
}
\arguments{
\item{asv}{Vector of sequences to be BLASTed.}

\item{db_path}{Complete path do formatted BLAST database.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}

\item{num_threads}{Number of threads to run BLAST on.
Passed on to BLAST+ argument \code{-num_threads}.}

\item{blast_type}{One of the available BLAST+ search engines,
#'   one of: \code{c("blastn", "blastp", "blastx", "tblastn", "tblastx")}.}

\item{perc_id}{Lowest identity percentage cutoff.
Passed on to BLAST+ \code{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff.
Passed on to BLAST+ \code{-qcov_hsp_perc}.}

\item{verbose}{Should condathis::run() internal command be shown?}

<<<<<<< HEAD
\item{env_name}{The name of the conda environment used to run
command-line tools (i.e. "blastr-blast-env").}
=======
\item{env_name}{The name of the conda environment with the parameter (i.e. "blast-env")}

\item{gapopen}{Gap open penalty (default  = 5 for blastn, = 11 for other blast)}

\item{gapextend}{Gap extend penalty (default  = 2 for blastn, = 1 for other blast)}
>>>>>>> local update to lucio's version
}
\value{
Unformatted BLAST results.
For results formatted as tibble, please use \verb{[get_blast_results()]}
}
\description{
Run BLAST for a single sequence and return raw results.
For formatted results, use the function \verb{[BLASTr::get_blastn_results()]}.
}
\examples{
\dontrun{

dna_fasta_path <- fs::path_package("BLASTr", "extdata", "minimal_db_blast", ext = "fasta")
temp_db_path <- fs::path_temp("minimal_db_blast")
make_blast_db(fasta_path = dna_fasta_path, db_path = temp_db_path)
asvs_string <- "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAG
AGTACTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC"
blast_res <- run_blast(
  asv = asvs_string,
  db_path = temp_db_path
)
}

}
