% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run BLAST}
\usage{
run_blast(
  asv,
  db_path,
  num_alignments = 4,
  num_threads = 1,
  blast_type = "blastn",
  perc_id = 80,
  perc_qcov_hsp = 80
)
}
\arguments{
\item{db_path}{Complete path do formatted BLAST database.}

\item{num_alignments}{Number of alignments to retrieve from BLAST. Max = 6.}

\item{num_threads}{Number of threads to run BLAST on.
Passed on to BLAST+ argument \emph{-num_threads}.}

\item{blast_type}{One of the available BLAST+ search engines,
one of: "blastn", "blastp", "blastx", "tblastn", "tblastx".}

\item{perc_id}{Lowest identity percentage cutoff.
Passed on to BLAST+ \emph{-perc_identity}.}

\item{perc_qcov_hsp}{Lowest query coverage per HSP percentage cutoff.
Passed on to BLAST+ \emph{-qcov_hsp_perc}.}

\item{asvs}{Vector of sequences to be BLASTed.}
}
\value{
Unformatted BLAST results.
For results formatted as tibble, please use \code{BLASTr::get_blast_results()}
}
\description{
Run BLAST for a single sequence and return raw results.
For formated results, use the function \emph{BLASTr::getblastn_results()}
}
\examples{
blast_res <- BLASTr::run_blast(
  asv = "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAG
    AGTACTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC",
  db_path = "/data/databases/nt/nt",
  perc_id = 80,
  num_thread = 1,
  perc_qcov_hsp = 80,
  num_alignments = 2
)

}
