% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getBitscore}
\alias{getBitscore}
\alias{getEvalue}
\alias{getMaxBitscore}
\alias{getScore}
\alias{getTotalBitscore}
\title{Access components of High-Scoring Pairs (Hsps)}
\usage{
getBitscore(x, max = FALSE, sum = FALSE)

getMaxBitscore(x)

getTotalBitscore(x)

getScore(x, max = FALSE)

getEvalue(x, max = FALSE)
}
\arguments{
\item{x}{A \linkS4class{Hsp} or \linkS4class{HspList}.}

\item{max}{If \code{TRUE} return only the highest score.}

\item{sum}{If \code{TRUE} sum scores from all hsps of a Blast hit.}
}
\value{
An numeric vector or a list of numeric vectors.
}
\description{
Extract the bit score, raw score, or expect values.
}

