% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{class}
\name{sqliteDB-class}
\alias{.sqliteDB}
\alias{sqliteDB-class}
\title{sqliteDB-class}
\description{
sqliteDB-class
}
\section{Fields}{

\describe{
\item{\code{.con}}{An \code{\linkS4class{SQLiteConnection}}.}

\item{\code{.info}}{A list.}

\item{\code{.path}}{Path to the database.}
}}
\keyword{internal}

